### 2025-03-21 #7939 @editorialbot
[PRE REVIEW]: dpest: Streamlining Creation of PEST input files for DSSAT Crop Model Calibration
**Submitting author:** @DS4Ag (<a href="http://orcid.org/0000-0001-8610-9901">Luis Vargas Rojas</a>)
**Repository:** https://github.com/DS4Ag/dpest
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.3
