### 2025-03-21 #7939 @DS4Ag
The following researchers have reviewed a paper (https://joss.theoj.org/papers/10.21105/joss.05548) that describes a Python tool similar to dpest, which also creates PEST input files but for MIKE SHE, a hydrological modeling software. This tool shares similarities with dpest in its functionality, as both automate the creation of PEST input files for different models: **aleaf**
**ecomodeller**
