+++ paper/paper.bib
@@ -7,12 +7,20 @@
     isbn = {978-0-9943786-0-6}
 }
 
-@book{Hoogenboom:2023,
-    author = {{Hoogenboom}, G. and {Porter}, C.H. and {Shelia}, V. and {Boote}, K.J. and {Singh}, U. and {Pavan}, W. and {Oliveira}, F.A.A. and {Moreno-Cadena}, L.P. and {Ferreira}, T.B. and {White}, J.W. and {Lizaso}, J.I. and {Pequeno}, D.N.L. and {Kimball}, B.A. and {Alderman}, P.D. and {Thorp}, K.R. and {Cuadra}, S.V. and {Vianna}, M.S. and {Villalobos}, F.J. and {Batchelor}, W.D. and {Asseng}, S. and {Jones}, M.R. and {Hopf}, A. and {Dias}, H.B. and {Hunt}, L.A. and {Jones}, J.W.},
-    title = {Decision Support System for Agrotechnology Transfer (DSSAT) Version 4.8.2},
-    publisher = {DSSAT},
-    year = {2023},
-    url = {https://www.DSSAT.net}
+@article{JONES:2003,
+  author    = {J. W. Jones and G. Hoogenboom and C. H. Porter and K. J. Boote and W. D. Batchelor and L. A. Hunt and P. W. Wilkens and U. Singh and A. J. Gijsman and J. T. Ritchie},
+  title     = {The {DSSAT} cropping system model},
+  journal   = {European Journal of Agronomy},
+  volume    = {18},
+  number    = {3},
+  pages     = {235-265},
+  year      = {2003},
+  note      = {Modelling Cropping Systems: Science, Software and Applications},
+  issn      = {1161-0301},
+  doi       = {10.1016/S1161-0301(02)00107-7},
+  url       = {https://www.sciencedirect.com/science/article/pii/S1161030102001077},
+  keywords  = {Crop simulation, Weather, Research tool, Decision aid},
+  abstract  = {The decision support system for agrotechnology transfer (DSSAT) has been in use for the last 15 years by researchers worldwide. This package incorporates models of 16 different crops with software that facilitates the evaluation and application of the crop models for different purposes. Over the last few years, it has become increasingly difficult to maintain the DSSAT crop models, partly due to the fact that there were different sets of computer code for different crops with little attention to software design at the level of crop models themselves. Thus, the DSSAT crop models have been re-designed and programmed to facilitate more efficient incorporation of new scientific advances, applications, documentation and maintenance. The basis for the new DSSAT cropping system model (CSM) design is a modular structure in which components separate along scientific discipline lines and are structured to allow easy replacement or addition of modules. It has one Soil module, a Crop Template module which can simulate different crops by defining species input files, an interface to add individual crop models if they have the same design and interface, a Weather module, and a module for dealing with competition for light and water among the soil, plants, and atmosphere. It is also designed for incorporation into various application packages, ranging from those that help researchers adapt and test the CSM to those that operate the DSSAT--CSM to simulate production over time and space for different purposes. In this paper, we describe this new DSSAT--CSM design as well as approaches used to model the primary scientific components (soil, crop, weather, and management). In addition, the paper describes data requirements and methods used for model evaluation. We provide an overview of the hundreds of published studies in which the DSSAT crop models have been used for various applications. The benefits of the new, re-designed DSSAT--CSM will provide considerable opportunities to its developers and others in the scientific community for greater cooperation in interdisciplinary research and in the application of knowledge to solve problems at field, farm, and higher levels.}
 }
 
 @article{Kasampalis:2018,
