+++ docs/example.rst
@@ -18,7 +18,7 @@ This example demonstrates how to use ``dpest`` to create the necessary files for
 
 .. raw:: html
 
-    <iframe width="560" height="315" 
+    <iframe 
             src="https://www.youtube.com/embed/dzKpvJSEXZc?vq=hd1080" 
             frameborder="0" allowfullscreen>
     </iframe>
