+++ dpest/wheat/overview.py
@@ -5,7 +5,8 @@ def overview(
     treatment = None,
     overview_file_path = None,
     output_path = None,
-    variable_classifications = None,
+    variables = None,
+    variables_classification = None,
     overview_ins_first_line = None,
     mrk = '~',
     smk = '!',
@@ -26,7 +27,8 @@ def overview(
     =======
 
         * **output_path** (*str*, *default: current working directory*): Directory where the generated ``PEST instruction file (.INS)`` will be saved.
-        * **variable_classifications** (*dict*): Mapping of variable names to their respective categories. If not provided, defaults to a pre-configured classification scheme defined in the package. Users can override this by providing their own dictionary to define the variables from the *MAIN GROWTH AND DEVELOPMENT VARIABLES section of the ``OVERVIEW.OUT`` DSSAT file, using the format ``{variable: variable_group, …}``. Variables group names should be less than 12 characters.
+        * **variables** (*list* or *str*): Variable(s) from the ``OVERVIEW.OUT`` file that PEST will extract in case the user does not want to use all the variables present in the DSSAT “A file” for the calibration. The PEST instruction file will use these to read the model output. You may specify a single variable as a string (e.g., 'Anthesis (DAP)') or multiple variables as a list (e.g., ['Anthesis (DAP)', 'Maturity (DAP)', 'Product wt (kg dm/ha;no loss)', 'Maximum leaf area index',  'Canopy (tops) wt (kg dm/ha)', 'Above-ground N (kg/ha)']).
+        * **variables_classification** (*dict*): Mapping of variable names to their respective categories. If not provided, defaults to a pre-configured classification scheme defined in the package. Users can override this by providing their own dictionary to define the variables from the *MAIN GROWTH AND DEVELOPMENT VARIABLES section of the ``OVERVIEW.OUT`` DSSAT file, using the format ``{variable: variable_group, …}``. Variables group names should be less than 12 characters.
         * **overview_ins_first_line** (*str*, *default: "pif"*): First line of the ``PEST instruction file (.INS)``. This is the PEST default value and should not be changed without good reason.
         * **mrk** (*str*, *default: "~"*) Primary marker delimiter character for the instruction file. Must be a single character and cannot be A-Z, a-z, 0-9, !, [, ], (, ), :, space, tab, or &.
         * **smk** (*str*, *default: "!"*) Secondary marker delimiter character for the instruction file. Must be a single character and cannot be A-Z, a-z, 0-9, [, ], (, ), :, space, tab, or &.
@@ -67,7 +69,14 @@ def overview(
           overview(
               treatment = '164.0 KG N/HA DRY',
               overview_file_path = 'C:/DSSAT48/Wheat/OVERVIEW.OUT',
-              variable_classifications = {
+              variables = [
+                  'Anthesis (DAP)', 'Maturity (DAP)',
+                  'Product wt (kg dm/ha;no loss)',
+                  'Maximum leaf area index',
+                  'Canopy (tops) wt (kg dm/ha)',
+                  'Above-ground N (kg/ha)'
+              ]
+              variables_classification = {
                   'Anthesis (DAP)': 'phenology',
                   'Maturity (DAP)': 'phenology',
                   'Product wt (kg dm/ha;no loss)': 'yield',
@@ -77,13 +86,13 @@ def overview(
               }
           )
 
-       This example demonstrates how to specify the ``variable_classifications`` argument to group specific variables. In this case, the returned tuple is not saved, but the ``PEST instruction file (.INS)`` is still created at the specified location. If you want to use the cultivar parameters and path for the ``pst`` module, the returned tuple should be saved in two variables. Additionally, the example shows how the  ``variable_classifications`` optional variable should be entered as a dictionary.
[…]
