+++ dpest/wheat/plantgro.py
@@ -6,7 +6,7 @@ def plantgro(
     treatment = None, 
     variables = None, 
     output_path = None,
-    variable_classifications = None,
+    variables_classification = None,
     plantgro_ins_first_line = None,
     mrk = '~',
     smk = '!',
@@ -25,7 +25,7 @@ def plantgro(
     =======
 
         * **output_path** (*str*, *default: current working directory*): Directory where the generated ``PEST instruction file (.INS)`` will be saved.
-        * **variable_classifications** (*dict*): Mapping of ``variable`` names to their respective categories. If not provided, defaults to a pre-configured classification scheme defined in the package. Users can override this by providing their own dictionary in the format ``{variable: variable_group, …}``. Variables group names should be less than 12 characters.
+        * **variables_classification** (*dict*): Mapping of ``variable`` names to their respective categories. If not provided, defaults to a pre-configured classification scheme defined in the package. Users can override this by providing their own dictionary in the format ``{variable: variable_group, …}``. Variables group names should be less than 12 characters.
         * **plantgro_ins_first_line** (*str*, *default: "pif"*): First line of the ``PEST instruction file (.INS)``. This is the PEST default value and should not be changed without good reason.
         * **mrk** (*str*, *default: "~"*): Primary marker delimiter character for the instruction file. Must be a single character and cannot be A-Z, a-z, 0-9, !, [, ], (, ), :, space, tab, or &.
         * **smk** (*str*, *default: "!"*): Secondary marker delimiter character for the instruction file. Must be a single character and cannot be A-Z, a-z, 0-9, [, ], (, ), :, space, tab, or &.
@@ -68,7 +68,7 @@ def plantgro(
               plantgro_file_path = 'C:/DSSAT48/Wheat/PlantGro.OUT',
               treatment = '164.0 KG N/HA IRRIG',
               variables = 'LAID',
-              variable_classifications = {
+              variables_classification = {
                   'LAID': 'lai'
               }
           )
@@ -79,7 +79,7 @@ def plantgro(
     # Define default variables:
     yml_file_block = 'PLANTGRO_FILE'
     yaml_file_variables = 'INS_FILE_VARIABLES'
-    yaml_variable_classifications = 'VARIABLE_CLASSIFICATIONS'
+    yaml_variables_classification = 'VARIABLES_CLASSIFICATION'
 
     try:
         ## Get the yaml_data
@@ -93,9 +93,6 @@ def plantgro(
         # Load YAML configuration
         with open(arguments_file, 'r') as yml_file:
             yaml_data = yaml.safe_load(yml_file)
-
-        # Validate plantgro_file_path
-        validate_file(plantgro_file_path, '.OUT')
 
         # Validate treatment
         if not treatment or not isinstance(treatment, str):
@@ -121,23 +118,26 @@ def plantgro(
         if mrk == smk:
             raise ValueError("mrk and smk must be different characters.")
 
-        # Validate variable_classifications
-        if variable_classifications is None:
-            variable_classifications = yaml_data[yml_file_block][yaml_variable_classifications]
+        # Validate variables_classification
+        if variables_classification is None:
+            variables_classification = yaml_data[yml_file_block][yaml_variables_classification]
 
         if plantgro_ins_first_line is None:
             # Load default arguments from the YAML file if not provided
             function_arguments = yaml_data[yml_file_block][yaml_file_variables]
             plantgro_ins_first_line = function_arguments['first_line']
 
+        # Validate plantgro_file_path
+        validated_path = validate_file(plantgro_file_path, '.OUT')
+
         # Get treatment number
-        treatment_dict = simulations_lines(plantgro_file_path)
[…]
