### 2025-04-01 af7ad98
Add variables argument to the overview module
