+++ pyproject.toml
@@ -9,7 +9,7 @@ requires-python = ">=3.7"
 classifiers = [
     "Programming Language :: Python :: 3",
     "Operating System :: OS Independent",
-    "License :: OSI Approved :: GNU General Public License v3.0-only"
+    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)"
 ]
 
 dependencies = [
