### 2025-04-02 9f97db1
extract_simulation_data function column names simulated and measured updated
