### 2025-04-14 4a03314
Suffix arg for multiple treatments
