### 2025-04-19 e05691d
Function unoptmax to update NOPTMAX .pst value
