+++ docs/index.rst
@@ -40,12 +40,16 @@ Table of Contents
    example_multiple_trts
 
 *  `Basic Usage <basic_usage.html>`_
-*  Module Reference
+*  Create the imput files using the ceres wheat model
     *  `dpest.wheat.ceres.cul <dpest.wheat.ceres.cul.html>`_
     *  `dpest.wheat.ceres.eco <dpest.wheat.ceres.eco.html>`_
     *  `dpest.wheat.overview <dpest.wheat.overview.html>`_
     *  `dpest.wheat.plantgro <dpest.wheat.plantgro.html>`_
+    * Utils 
+      *  `dpest.wheat.utils.uplantgro <dpest.wheat.utils.uplantgro.html>`_
+*  Create the PEST control file
     *  `dpest.pst <dpest.pst.html>`_
-    *  `dpest.wheat.utils.uplantgro <dpest.wheat.utils.uplantgro.html>`_
+    * Utils 
+      *  `dpest.pst <dpest.utils.unoptmax.html>`_
 *  `Example: Calibrating DSSAT for Wheat (CERES Model) <example.html>`_
 *  `Example: Calibrating a Cultivar Using Multiple Treatments <example_multiple_trts.html>`_
