### 2025-04-19 1425b2a
Function unoptmax to update NOPTMAX .pst value
