### 2025-04-19 613a987
Function unoptmax to update NOPTMAX .pst value
