### 2025-04-19 4e08a5a
Function unoptmax to update NOPTMAX .pst value
