### 2025-04-19 bb97576
Function unoptmax to update NOPTMAX .pst value
