### 2025-04-19 9cd0ee9
Function unoptmax to update NOPTMAX .pst value
