### 2025-04-22 #7939 @kthyng
Hi @DS4Ag! You are toward the end of 14 submissions waiting to be processed. I am processing approximately as we have editor capacity available, but editors are also chosen based on their expertise area so it isn't a linear change in time as editors become available.

That said, there is a comment above about addressing any potentially missing DOIs so you could work on that in the meantime.
