### 2025-04-22 #7939 @DS4Ag
Hi @kthyng, thank you for the update and for managing the high volume of submissions so thoughtfully.

Regarding the DOI check, we have reviewed the bibliography and can confirm that all valid DOIs have been included where available. The lagged items refer to resources that do not have DOIs, specifically a book chapter and a conference abstract. These were cited due to their relevance and importance for documenting PEST and crop model calibration using **dpest**, even though they don’t have registered DOIs.

We are grateful for your work in shepherding the process, and we stand by to support the next steps once an editor becomes available.
