+++ dpest/utils/unoptmax.py
@@ -4,8 +4,7 @@ def unoptmax(pst_path, new_value=0):
 
     The NOPTMAX parameter specifies the number of optimization iterations or model runs that PEST will perform.
     This function allows you to set NOPTMAX to any integer value, such as 0 (run the model once, no optimization)
-    or a large value for iterative optimization. The rest of the control file, including solver settings (e.g., LSQR),
-    is preserved.
+    or a large value for iterative optimization.
 
     **Required Arguments:**
     =======
@@ -46,25 +45,43 @@ def unoptmax(pst_path, new_value=0):
           pst_file_path = './ENTRY1/PEST_CONTROL.pst'
           unoptmax(pst_file_path, new_value = 10000)
     """
-    with open(pst_path, 'r') as f:
-        lines = f.readlines()
+    try:
+        # Validation for NOPTMAX value
+        if not isinstance(new_value, int):
+            raise ValueError(f"NOPTMAX must be an integer. Got: {new_value}")
+        if new_value not in [-2, -1, 0] and new_value <= 0:
+            raise ValueError("NOPTMAX must be -2, -1, 0, or any integer greater than zero.")
 
-    # NOPTMAX is on line 9 (index 8) in standard PEST control files
-    target_line_idx = 8
-    current_line = lines[target_line_idx]
+        with open(pst_path, 'r') as f:
+            lines = f.readlines()
 
-    # Split the line by whitespace to get all values
-    values = current_line.split()
+        # NOPTMAX is on line 9 (index 8) in standard PEST control files
+        target_line_idx = 8
+        if target_line_idx >= len(lines):
+            raise IndexError(f"Expected at least {target_line_idx + 1} lines in the file, but got {len(lines)}.")
 
-    # Replace first value with new_value, preserving alignment
-    # Calculate padding based on current value's width
-    current_padding = len(current_line) - len(current_line.lstrip())
-    formatted_value = f"{new_value:4d}"  # Right-aligned in 4 spaces
+        current_line = lines[target_line_idx]
+        values = current_line.split()
 
-    # Reconstruct the line, preserving the rest of the values
-    new_line = " " * current_padding + formatted_value + "   " + "   ".join(values[1:]) + "\n"
+        if not values:
+            raise ValueError("NOPTMAX line not found or is empty in the control file.")
 
-    lines[target_line_idx] = new_line
+        # Replace first value with new_value, preserving alignment
+        current_padding = len(current_line) - len(current_line.lstrip())
+        formatted_value = f"{new_value:d}"
 
-    with open(pst_path, 'w') as f:
-        f.writelines(lines)
+        # Reconstruct the line, preserving the rest of the values
+        new_line = " " * current_padding + formatted_value + "   " + "   ".join(values[1:]) + "\n"
+        lines[target_line_idx] = new_line
+
+        with open(pst_path, 'w') as f:
+            f.writelines(lines)
+
+    except FileNotFoundError:
+        print(f"Error: The file '{pst_path}' was not found.")
+    except IndexError as ie:
+        print(f"IndexError: {ie}")
+    except ValueError as ve:
+        print(f"ValueError: {ve}")
+    except Exception as e:
+        print(f"An unexpected error occurred: {e}")
