### 2025-04-24 #7939 @elbeejay
@ecomodeller, @aleaf, @tenemark would any of you be interested in reviewing this submission to JOSS titled: "dpest: Streamlining Creation of PEST input files for DSSAT Crop Model Calibration"?

I know you all have experience with JOSS in the past on a similar submission, related to the `mikecalsetup` Python package. The process at JOSS has not changed, we aim for an open and transparent review process via GitHub, and ask that our reviewers try to complete their intial reviews of the submission within about 6 weeks of the review formally beginning.

Please let me know if you're able to help out and review for JOSS!
