### 2025-04-24 444cdd0
Functions to modify control data section
