+++ docs/dpest.utils.unoptmax.rst
@@ -1,7 +1,0 @@
-dpest.utils.unoptmax
-----------------------------------
-
-.. automodule:: dpest.utils.unoptmax
-   :members:
-   :undoc-members:
-   :show-inheritance:
