### 2025-04-24 afc8052
Functions to modify control data section
