+++ docs/index.rst
@@ -35,11 +35,11 @@ Table of Contents
    :maxdepth: 2
    :caption: PEST control file
 
-   .. toctree::
-      :maxdepth: 2
-      :caption: API Reference
+.. toctree::
+   :maxdepth: 2
+   :caption: Control File Update Utilities 
 
-      utils
+   dpest.utils
 
 .. toctree::
    :maxdepth: 2
