### 2025-04-24 8a097d7
Functions to modify control data section
