+++ docs/utils.rst
@@ -1,29 +1,9 @@
-Utilities
-=========
+dpest.utils
+===========
+
+This module contains utility functions to modify the control data section of PEST input files.
 
 .. automodule:: dpest.utils
    :members:
    :undoc-members:
    :show-inheritance:
-
-.. toctree::
-   :hidden:
-
-   dpest.utils.unoptmax
-   dpest.utils.ufacorig
-   dpest.utils.ufacparmax
-   dpest.utils.unphinored
-   dpest.utils.unphistp
-   dpest.utils.unrelpar
-   dpest.utils.unumlam
-   dpest.utils.upestmode
-   dpest.utils.uphiratsuf
-   dpest.utils.uphiredlam
-   dpest.utils.uphiredstp
-   dpest.utils.uphiredswh
-   dpest.utils.urelparmax
-   dpest.utils.urelparstp
-   dpest.utils.urlambda1
-   dpest.utils.urlamfac
-   dpest.utils.urstfle
-   dpest.utils.rmv_splitcols
