### 2025-04-24 6dd5175
Functions to modify control data section
