+++ docs/utils.rst
@@ -1,9 +1,28 @@
-dpest.utils
+Control File Update Utilities
 ===========
 
-This module contains utility functions to modify the control data section of PEST input files.
+This subpackage contains utility functions to modify the control data section of PEST input files.
 
-.. automodule:: dpest.utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
+.. autosummary::
+   :toctree: utils_modules
+   :recursive:
+   :template: module.rst
+
+   dpest.utils.urstfle
+   dpest.utils.upestmode
+   dpest.utils.urlambda1
+   dpest.utils.urlamfac
+   dpest.utils.uphiratsuf
+   dpest.utils.uphiredlam
+   dpest.utils.unumlam
+   dpest.utils.urelparmax
+   dpest.utils.ufacparmax
+   dpest.utils.ufacorig
+   dpest.utils.uphiredswh
+   dpest.utils.unoptmax
+   dpest.utils.uphiredstp
+   dpest.utils.unphistp
+   dpest.utils.unphinored
+   dpest.utils.urelparstp
+   dpest.utils.unrelpar
+   dpest.utils.rmv_splitcols
