### 2025-04-24 1afa389
Functions to modify control data section
