+++ README.md
@@ -40,7 +40,7 @@ pip install dpest
 
 
 The following steps provide a brief overview of how to use `dpest`. For a **detailed, step-by-step example**, please refer to the official documentation:  
-👉 [Complete Example on Read the Docs](https://dpest.readthedocs.io/en/latest/example.html)  
+[Complete Example on Read the Docs](https://dpest.readthedocs.io/en/latest/example.html)  
 
 1.  **Locate DSSAT Genotype Files:** The cultivar (`.CUL`) and ecotype (`.ECO`) files are typically included with the DSSAT installation. These are usually found in the `C:\DSSAT48\Genotype\` directory.
 2.  **Run a DSSAT Simulation:** Execute a DSSAT simulation for your chosen wheat experiment using the CERES model. This will generate the necessary output files (`OVERVIEW.OUT` and `PlantGro.OUT`), typically found in the `C:\DSSAT48\Wheat\` directory.
@@ -133,6 +133,44 @@ dpest.pst(
 ```
 
 7.  **Run PEST:** Calibrate the model using PEST.
+
+---
+## Utilities for Editing PEST Control Files
+
+In addition to automating the creation of PEST control and template files for DSSAT CERES-Wheat model calibration, `dpest` includes a growing set of utility functions under `dpest.utils` for directly modifying existing `.pst` (PEST control) files.
+
+
+#### Why use `dpest.utils`?
+
+Unlike libraries such as [`pyEMU`](https://pyemu.readthedocs.io/en/develop/pyemu.pst.html), which parse `.pst` files into Python objects and rewrite the entire file when saving any change, `dpest.utils` performs lightweight, line-by-line edits to preserve the original structure and untouched content.
+
+These utilities:
+
+- Edit existing `.pst` files in place, without reconstructing or redefining unrelated fields.
+-  Are model-agnostic: they work with any `.pst` file, not just DSSAT-related ones.
+- Only modify model-independent control parameters (e.g., optimization settings), leaving parameter and data blocks untouched.
+
+This makes them ideal for **quickly tuning optimization settings** or **cleaning up `.pst` files** generated by `dpest` or other tools.
+
+#### Example Usage
+
+```python
+# Load the dpest.utils
+from dpest.utils import unoptmax, rmv_splitcols
+
+# Path to the .pst file
+pst_file_path = './ENTRY1/PEST_CONTROL.pst'
+
+# Increase the number of optimization iterations (NOPTMAX) in a .pst file
+unoptmax(pst_file_path, new_value = 10000)
+
+# Remove SPLITTHRESH/SPLITRELDIFF/SPLITACTION columns from a .pst file parameter groups section
+rmv_splitcols(“PEST_CONTROL.pst”)
+
+```
+
+**Full List of Utility Functions:**  
+For the complete reference of available utility functions and their descriptions, see the [dpest ReadTheDocs Utils page](https://dpest.readthedocs.io/en/latest/utils.html).
 
 ### Contributing
 
