### 2025-04-24 cb820c3
Functions to modify control data section
