### 2025-04-25 1b4ced6
Functions to modify control data section
