+++ dpest/utils/uphiredstp.py
@@ -18,6 +18,7 @@ def uphiredstp(pst_path, new_value):
 
     **Examples:**
     =======
+    
     .. code-block:: python
 
         from dpest.utils import uphiredstp
