+++ dpest/utils/uphiredswh.py
@@ -18,6 +18,7 @@ def uphiredswh(pst_path, new_value):
 
     **Examples:**
     =======
+    
     .. code-block:: python
 
         from dpest.utils import uphiredswh
