### 2025-04-25 eabee21
Functions to modify control data section
