### 2025-04-25 e3d8fea
Utils description added to paper
