+++ paper/paper.md
@@ -27,13 +27,13 @@ license: GPLv3+
 
 Process-based crop models simulate plant growth and can support crop improvement research by enacting different “what-if” scenarios. Decision Support System for Agrotechnology Transfer (DSSAT) is one of the most commonly used crop modeling platforms, containing modules that can be used to model different species [@JONES:2003]. To simulate different crop varieties, DSSAT calibration is needed. This involves adjusting model parameters, which are values stored in a model input file, to determine which sets of values give rise to simulations that most closely match measured data from field experiments; these data can include single timepoint measurements as well as time-series information. 
 
-While the DSSAT installation includes Graphical User Interface (GUI) tools specifically designed for calibration, they are not amenable to time-series data. In contrast, the Parameter Estimation and Uncertainty Analysis (PEST), a Command Line Interface (CLI) software, is a model-independent tool [@Doherty:2015] that can calibrate DSSAT models with high computational speed using not only the end-of-season crop data but also measurements collected within-season. To carry out model calibration, PEST requires three types of input files: template files (.TPL) that specify the parameters to calibrate, instruction files (.INS) that contain the guidelines to extract the model outputs, and the control file (.PST) that includes the specifications to manage all the calibration settings. Generation of these input files for PEST-based DSSAT calibration, however, can be a complex and time-consuming process, making it difficult to scale, for example, across many cultivars or strains. 	
+While the DSSAT installation includes Graphical User Interface (GUI) tools specifically designed for calibration, they are not amenable to time-series data. In contrast, the Parameter Estimation and Uncertainty Analysis (PEST), a Command Line Interface (CLI) software, is a model-independent tool [@Doherty:2015] that can calibrate DSSAT models with high computational speed using not only the end-of-season crop data but also measurements collected within-season. To carry out model calibration, PEST requires three types of input files: template files (.TPL) that specify the parameters to calibrate, instruction files (.INS) that contain the guidelines to extract the model outputs, and the control file (.PST) that includes the specifications to manage all the calibration settings. Generation of these input files for PEST-based DSSAT calibration, however, can be a complex and time-consuming process, making it difficult to scale, for example, across many cultivars or strains.   
 
 To address these challenges, we developed *dpest*, a Python library that streamlines the generation of PEST input files for calibrating DSSAT; the current version supports DSSAT’s wheat models. Importantly, using dpest enables researchers to script the entire calibration process, thereby enhancing the efficiency of the workflow to scale easily across multiple varieties. 
 
 ## Statement of Need
 
-With the adoption of new data collection technologies in agriculture, plant researchers have been developing methodologies to integrate remote sensing data and crop models to simulate crop performance and development [@Kasampalis:2018]. Remote sensing data have many advantages over traditional data collection methods in crop science. For instance, they can be collected repeatedly over time without disturbing the plants and at a low cost. Using these kinds of time-series data can improve the accuracy of crop model calibration. However, the calibration tools included in the DSSAT installation do not support the use of time-series data, which limits the potential of remote sensing for model calibration. To address this limitation, researchers can use PEST-based calibrations for the DSSAT models. However, the manual preparation of PEST input files requires expertise in both the PEST software and DSSAT models, which makes it a complex and time-consuming task. The *dpest* Python library helps overcome these challenges, allowing crop researchers to integrate time-series data from either remote sensing or direct measurements to improve PEST-based DSSAT model calibration. Moreover, streamlining the calibration process using *dpest* can facilitate the application of DSSAT models to large populations of crop varieties. 
[…]
