### 2025-04-25 6ec85a7
lsqr function
