### 2025-04-25 3b613fa
lsqr function
