+++ dpest/utils/lsqr.py
@@ -1,0 +1,214 @@
+def ulqsr(pst_path, lsqrmode=None, lsqr_atol=None, lsqr_btol=None,
+          lsqr_conlim=None, lsqr_itnlim=None, lsqrwrite=None):
+    """
+    Adds or updates the LSQR section in a PEST control (.pst) file.
+
+    The LSQR section configures PEST to use the LSQR algorithm for solving the inverse problem.
+    This is especially useful for large models with many parameters.
+
+    **Required Arguments:**
+    =======
+        * **pst_path** (*str*):
+            Path to the .pst PEST control file to modify.
+
+    **Optional Arguments:**
+    =======
+        * **lsqrmode** (*int*):
+            LSQR mode (must be 0 or 1).
+            Default is 1 (enable LSQR).
+        * **lsqr_atol** (*float*):
+            LSQR algorithm atol variable (must be ≥ 0).
+            Default is 1e-4.
+        * **lsqr_btol** (*float*):
+            LSQR algorithm btol variable (must be ≥ 0).
+            Default is 1e-4.
+        * **lsqr_conlim** (*float*):
+            LSQR algorithm conlim variable (must be ≥ 0).
+            Default is 28.0.
+        * **lsqr_itnlim** (*int*):
+            LSQR algorithm itnlim variable (must be > 0).
+            Default is 28.
+        * **lsqrwrite** (*int*):
+            Write LSQR file flag (must be 0 or 1).
+            Default is 0 (don't write LSQR file).
+
+    **Returns:**
+    =======
+        * ``None``
+
+    **Examples:**
+    =======
+    
+    1. **Adding an LSQR Section to a PEST Control File with Default Parameters:**
+
+       .. code-block:: python
+
+          from dpest.utils import ulqsr
+
+          ulqsr(
+              pst_path = "PEST_CONTROL.pst"
+          )
+
+       This example demonstrates adding a new LSQR section to a PEST control file using all default parameter values (lsqrmode=1, lsqr_atol=1e-4, lsqr_btol=1e-4, lsqr_conlim=28.0, lsqr_itnlim=28, lsqrwrite=0). The function will insert the LSQR section after any existing SVD section or append it to the file.
+
+    2. **Updating Specific LSQR Parameters in an Existing PEST Control File:**
+
+       .. code-block:: python
+
+          from dpest.utils import ulqsr
+
+          ulqsr(
+              pst_path = "PEST_CONTROL.pst",
+              lsqr_atol = 1e-6,
+              lsqr_btol = 1e-6,
+              lsqr_itnlim = 50
+          )
+
+       This example updates only the specified LSQR parameters (lsqr_atol, lsqr_btol, and lsqr_itnlim) while preserving existing values for other LSQR parameters. The function will modify the existing LSQR section if present, or create a new one with default values for unspecified parameters.
+
+    3. **Disabling LSQR Mode While Maintaining Other Settings:**
+
+       .. code-block:: python
+
+          from dpest.utils import ulqsr
+
+          ulqsr(
+              pst_path = "PEST_CONTROL.pst",
+              lsqrmode = 0
+          )
+
+       This example disables LSQR mode (lsqrmode=0) while keeping other LSQR parameters at their current values. If no LSQR section exists, it will be created with disabled mode and default values for other parameters.
+    """
+    try:
+        import os
+
+        # Default values for LSQR parameters
+        defaults = {
+            "lsqrmode": 1,
+            "lsqr_atol": 1e-4,
+            "lsqr_btol": 1e-4,
+            "lsqr_conlim": 28.0,
+            "lsqr_itnlim": 28,
+            "lsqrwrite": 0
+        }
+
+        # Verify file exists
[…]
