+++ dpest/utils/__init__.py
@@ -21,3 +21,5 @@ from .relparstp import relparstp
 from .nrelpar import nrelpar
 
 from .rmv_splitcols import rmv_splitcols
+
+from .lsqr import lsqr
