### 2025-04-26 3641536
lsqr and svd functions
