### 2025-04-27 6645e56
lsqr and svd functions
