+++ docs/utils.rst
@@ -28,4 +28,6 @@ This subpackage contains utility functions to modify the control data section of
    dpest.utils.nrelpar
    dpest.utils.lsqr.rst
    dpest.utils.svd.rst
+   dpest.utils.rmv_lsqr
+   dpest.utils.rmv_svd
    dpest.utils.rmv_splitcols
