### 2025-05-07 #7939 @DS4Ag
Hi @elbeejay, thanks for your work on this submission. I noticed you’ve recently invited reviewers, and I’d like to suggest two additional names who may be a good fit based on their past JOSS reviews related to crop and plant modeling:

* **dlebauer**, who reviewed *climatrends* (R package for crop and ecological modeling) and *agweather-qaqc* (Python package for agricultural weather QA/QC)
* **ashiklom**, who reviewed *PlantSimEngine* (a soil-plant-atmosphere simulation engine in Julia) and *prospect* (R package for modeling leaf optical properties)

Both seem to have relevant expertise in agricultural and plant modeling software and might provide valuable feedback.
