### 2025-05-10 #7939 @elbeejay
Hi @cwdunkerly, was wondering if you'd considering reviewing this submission to JOSS titled: "dpest: Streamlining Creation of PEST input files for DSSAT Crop Model Calibration"?

I know you previously published your `agweather-qaqc` package via JOSS, and I thought you might be an appropriate reviewer for this one. The process at JOSS has not changed since your last go-round, we aim for an open and transparent review process via GitHub, and ask that our reviewers try to complete their intial reviews of the submission within about 6 weeks of the review formally beginning.

Please let me know if you're able to help out and review for JOSS!
