### 2025-05-10 #7939 @julienmalard
@elbeejay I would be happy to review
