### 2025-05-10 #8188 @editorialbot
[REVIEW]: dpest: Streamlining Creation of PEST input files for DSSAT Crop Model Calibration
**Submitting author:** @DS4Ag (<a href="http://orcid.org/0000-0001-8610-9901">Luis Vargas Rojas</a>)
**Repository:** https://github.com/DS4Ag/dpest
**Branch with paper.md** (empty if default branch):
**Version:** v2.1.2
