### 2025-05-10 #8188 @elbeejay
@aleaf and @julienmalard, thank you both for volunteering your time and reviewing for JOSS. As mentioned in the pre-review issue, we are asking our reviewers to complete their first pass through the submission in 6 weeks, so by June 21, in this case. If you need more time, I just ask that you communicate that with us here.

The top post in this issue thread will have a link to the JOSS reviewer guidelines as well as instructions on how to generate your reviewer checklist. As a reminder, we strive for transparent and interactive reviews, feel free to make comments and suggestions as you review the submission - you don't have to wait until you've done your full review to start conversations with the authors about various aspects of the software.

If you have any questions or need anything, please feel free to ask me.

Thanks again,
Jay
