### 2025-05-16 #8188 @julienmalard
Hello,
Here is my initial review:
- Include link to documentation in the article
- The repository does not seem to include automated tests.
- Community guidelines seem to be missing
- Consider briefly citing similar DSSAT wrapper packages in different programming languages, even if they don't specifically support PEST files.
- The top menu bar of the documentation is a bit confusing (and redundant with the sidebar). In addition, the fact that `dpest` occurs below `table of contents` and `basic usage` in the sidebar but actually links to a header above the table of contents in the main page is a bit confusing at first.
