+++ docs/conf.py
@@ -35,6 +35,8 @@ html_theme = 'pydata_sphinx_theme'
 
 html_show_sourcelink = True
 
+html_title = "Home"
+
 # html_sidebars = {
 #     "**": ["sidebar-nav-bs"]
 # }
@@ -54,5 +56,5 @@ html_theme_options = {
     "use_source_button": True,
     "show_nav_level": 2,  # Ensure navigation is visible up to depth 2
     "navigation_with_keys": True,  # Enable keyboard navigation
-    "navbar_center": [],
+    "navbar_end": [],
 }
