+++ docs/conf.py
@@ -54,8 +54,10 @@ html_theme_options = {
     "github_url": "https://github.com/DS4Ag/dpest/",
     "use_repository_button": True,
     "use_source_button": True,
-    "show_nav_level": 2,  # Ensure navigation is visible up to depth 2
-    "navigation_with_keys": True,  # Enable keyboard navigation
+    "show_nav_level": 2,
+    "navigation_with_keys": True,
+    "navbar_start": [],
     "navbar_center": [],
-    "navbar_end": ["search-field.html"],  # <-- This puts the search box on the right
+    "navbar_end": ["search-field.html"],
+    "navbar_persistent": [],  # <--- This removes the default search button
 }
