+++ docs/conf.py
@@ -54,9 +54,7 @@ html_theme_options = {
     "github_url": "https://github.com/DS4Ag/dpest/",
     "use_repository_button": True,
     "use_source_button": True,
-    "show_nav_level": 2,
-    "navigation_with_keys": True,
+    "show_nav_level": 2,  # Ensure navigation is visible up to depth 2
+    "navigation_with_keys": True,  # Enable keyboard navigation
     "navbar_center": [],
-    "navbar_end": [],
-    "navbar_persistent": [],  # <--- This removes the default search button
 }
