+++ docs/conf.py
@@ -57,4 +57,5 @@ html_theme_options = {
     "show_nav_level": 2,  # Ensure navigation is visible up to depth 2
     "navigation_with_keys": True,  # Enable keyboard navigation
     "navbar_center": [],
+    "navbar_end": ["search-field.html", "navbar-icon-links"],
 }
