+++ docs/conf.py
@@ -56,5 +56,5 @@ html_theme_options = {
     "use_source_button": True,
     "show_nav_level": 2,  # Ensure navigation is visible up to depth 2
     "navigation_with_keys": True,  # Enable keyboard navigation
-    "navbar_center": ["<div style='flex: 1'></div>"],
+    "navbar_center": [],
 }
