+++ paper/paper.md
@@ -33,11 +33,11 @@ To address these challenges, we developed *dpest*, a Python library that streaml
 
 ## Statement of Need
 
-With the adoption of new data collection technologies in agriculture, plant researchers have been developing methodologies to integrate remote sensing data and crop models to simulate crop performance and development [@Kasampalis:2018]. Remote sensing data have many advantages over traditional data collection methods in crop science. For instance, they can be collected repeatedly over time without disturbing the plants and at a low cost. Using these kinds of time-series data can improve the accuracy of crop model calibration. However, the calibration tools included in the DSSAT installation do not support the use of time-series data, which limits the potential of remote sensing for model calibration. Researchers can use PEST-based calibrations for the DSSAT models to address this limitation, but manual preparation of PEST input files requires expertise in both the PEST software and DSSAT models, which makes it a complex and time-consuming task. The *dpest* Python library helps overcome these challenges, allowing crop researchers to integrate time-series data from either remote sensing or direct measurements to improve PEST-based DSSAT model calibration. Moreover, streamlining the calibration process using *dpest* can facilitate the application of DSSAT models to large populations of crop varieties. Finally,  *dpest* includes model-agnostic utilities for targeted modification of  PEST control files (.PST), reducing the need for full-file reconstruction and facilitating iterative calibration workflows.
+With the adoption of new data collection technologies in agriculture, plant researchers have been developing methodologies to integrate remote sensing data and crop models to simulate crop performance and development [@Kasampalis:2018]. Remote sensing data have many advantages over traditional data collection methods in crop science. For instance, they can be collected repeatedly over time without disturbing the plants and at a low cost. Using these kinds of time-series data can improve the accuracy of crop model calibration. However, the calibration tools included in the DSSAT installation do not support the use of time-series data, which limits the potential of remote sensing for model calibration. Researchers can use PEST-based calibrations for the DSSAT models to address this limitation, but manual preparation of PEST input files requires expertise in both the PEST software and DSSAT models, which makes it a complex and time-consuming task. Previous efforts to streamline this process include an R-based DSSAT-PEST script distributed as supplementary material with the paper [@Ma:2020], which automates PEST file generation but requires manual setup of configuration files. The *dpest* Python library helps overcome these challenges, allowing crop researchers to integrate time-series data from either remote sensing or direct measurements to improve PEST-based DSSAT model calibration. Moreover, streamlining the calibration process using *dpest* can facilitate the application of DSSAT models to large populations of crop varieties. Finally,  *dpest* includes model-agnostic utilities for targeted modification of  PEST control files (.PST), reducing the need for full-file reconstruction and facilitating iterative calibration workflows.
 
 ## Functionality
 
-The *dpest* package includes the following modules:
+The *dpest* package includes the following modules, each with detailed usage instructions and examples available in the [documentation](https://dpest.readthedocs.io/en/latest/):
 
  - *cul()*: Creates PEST template files (.TPL) for DSSAT cultivar parameters. The file is used for cultivar calibration.
  - *eco()*: Creates PEST template files (.TPL) for DSSAT ecotype parameters. The file is used for ecotype calibration.
