+++ paper/paper.bib
@@ -35,6 +35,18 @@
     url = {https://doi.org/10.3390/jimaging4040052}
 }
 
+@article{Ma:2020,
+    author = {{Ma}, H. and {Malone}, R. W. and {Jiang}, T. and {Yao}, N. and {Chen}, S. and {Song}, L. and {Feng}, H. and {Yu}, Q. and {He}, J.},
+    title = {Estimating crop genetic parameters for DSSAT with modified PEST software},
+    journal = {European Journal of Agronomy},
+    year = {2020},
+    volume = {115},
+    pages = {126017},
+    issn = {1161-0301},
+    doi = {10.1016/j.eja.2020.126017},
+    url = {https://www.sciencedirect.com/science/article/pii/S1161030120300253}
+}
+
 @misc{Vargas-Rojas:2024,
     author = {{Vargas-Rojas}, L. and {Wang}, D. and {Reynolds}, M.},
     title = {How Can We Streamline Wheat Model Calibration with Remote Sensing?},
@@ -44,3 +56,4 @@
     address = {San Antonio, TX},
     url = {https://scisoc.confex.com/scisoc/2024am/meetingapp.cgi/Paper/161215}
 }
+
