+++ .github/workflows/tests.yml
@@ -15,6 +15,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
+        pip install dpest
         pip install pytest
 
     - name: Run tests
