+++ tests/test_cul.py
@@ -6,8 +6,8 @@ import pytest
 def test_cul_template_generation(tmp_path):
     """Test generation of cultivar template files"""
     # Setup tests environment
-    test_data_dir = Path(__file__).parent / "test_data"  # Relative to test file
-    input_file = test_data_dir / "DSSAT48_data/Genotype/WHCER048.CUL"
+    test_data_dir = Path(__file__).parent / "DSSAT48_data"  # Relative to test file
+    input_file = test_data_dir / "Genotype/WHCER048.CUL"
     output_dir = tmp_path / "output"
 
     # Create tests output directory
