+++ tests/test_cul.py
@@ -1,7 +1,5 @@
 import dpest
 from pathlib import Path
-import pytest
-
 
 def test_cul_template_generation(tmp_path):
     """Test generation of cultivar template files."""
@@ -16,22 +14,21 @@ def test_cul_template_generation(tmp_path):
     # Create the output directory
     output_dir.mkdir(parents=True, exist_ok=True)
 
-    # Call the dpest.wheat.ceres.cul function and check return
+    # Call the dpest.wheat.ceres.cul function
     result = dpest.wheat.ceres.cul(
         P='P1D, P5',
         G='G1, G2, G3',
         PHINT='PHINT',
         cultivar='MANITOU',
-        cul_file_path=str(input_file),
-        output_path=str(output_dir)
+        cul_file_path=str(input_file),     # Ensure it's a string
+        output_path=str(output_dir)        # Ensure it's a string
     )
 
+    # Validate result
     assert result is not None, "Function returned None"
     assert isinstance(result, tuple) and len(result) == 2, "Unexpected return value format"
 
     params, tpl_path = result
 
-    # Validate the output template file
+    tpl_path = Path(tpl_path)  # Convert str to Path for file checking
     assert tpl_path.exists(), f"Template file not created: {tpl_path}"
-    # Optional: check contents or filename
-    # assert tpl_path.name == "WHCER048_CUL.TPL"
