+++ tests/test_cul.py
@@ -24,11 +24,22 @@ def test_cul_template_generation(tmp_path):
         output_path=str(output_dir)        # Ensure it's a string
     )
 
-    # Validate result
+    # 1. Validate result is not None
     assert result is not None, "Function returned None"
+
+    # 2. Validate result is a tuple with length 2
     assert isinstance(result, tuple) and len(result) == 2, "Unexpected return value format"
 
+    # 3. Unpack the result tuple
     params, tpl_path = result
 
+    # 4. Convert tpl_path to a Path and check that the file exists on disk.
     tpl_path = Path(tpl_path)  # Convert str to Path for file checking
     assert tpl_path.exists(), f"Template file not created: {tpl_path}"
+
+    # 5. Check that `params` is a dictionary
+    assert isinstance(params, dict), "Expected `params` to be a dictionary"
+
+    # 6. Check that the dictionary has the expected nested keys
+    expected_keys = {'parameters', 'minima_parameters', 'maxima_parameters', 'parameters_grouped'}
+    assert expected_keys.issubset(params), f"Missing expected keys in params: {expected_keys - set(params)}"
