+++ tests/test_eco.py
@@ -1,0 +1,44 @@
+import dpest
+from pathlib import Path
+
+def test_eco_template_generation(tmp_path):
+    """Test generation of ecotype template files."""
+    # Setup paths
+    repo_root = Path(__file__).parent.parent
+    input_file = repo_root / "tests/DSSAT48_data/Genotype/WHCER048.ECO"
+    output_dir = tmp_path / "output"
+
+    # Ensure the input file exists
+    assert input_file.exists(), f"Input file not found: {input_file}"
+
+    # Create the output directory
+    output_dir.mkdir(parents=True, exist_ok=True)
+
+    # Call the dpest.wheat.ceres.eco function
+    result = dpest.wheat.ceres.eco(
+        PHEN='P1, P2FR1',
+        VERN='VEFF',
+        ecotype='CAWH01',
+        eco_file_path=str(input_file),     # Ensure it's a string
+        output_path=str(output_dir)        # Ensure it's a string
+    )
+
+    # 1. Validate result is not None
+    assert result is not None, "Function returned None"
+
+    # 2. Validate result is a tuple with length 2
+    assert isinstance(result, tuple) and len(result) == 2, "Unexpected return value format"
+
+    # 3. Unpack the result tuple
+    params, tpl_path = result
+
+    # 4. Convert tpl_path to a Path and check that the file exists on disk.
+    tpl_path = Path(tpl_path)
+    assert tpl_path.exists(), f"Template file not created: {tpl_path}"
+
+    # 5. Check that `params` is a dictionary
+    assert isinstance(params, dict), "Expected `params` to be a dictionary"
+
+    # 6. Check that the dictionary has the expected nested keys
+    expected_keys = {'parameters', 'minima_parameters', 'maxima_parameters', 'parameters_grouped'}
+    assert expected_keys.issubset(params), f"Missing expected keys in params: {expected_keys - set(params)}"
