+++ examples/wheat/ceres/usage_example.py
@@ -1,361 +1,0 @@
-#!/usr/bin/env python
-# coding: utf-8
-
-# ## 1. Prerequisites for Using dpest
-# 
-# ### 1.1. Using Pre-generated DSSAT Files
-# 
-# *dpest* is designed to interface with DSSAT, a third-party crop modeling suite widely used for simulating wheat growth and yield using the **SWSW7501WH N RESPONSE experiment** included in the **CERES-Wheat model**. **Full use of dpest requires DSSAT (v4.8+), freely available at the [DSSAT website](https://dssat.net/).** To enable verification without DSSAT installation, this repository includes pre-generated DSSAT files in a standardized directory structure:
-# 
-# ```
-# DSSAT_data/  
-# ├── Genotype/          # Contains cultivar and ecotype parameter files  
-# │   ├── WHCER048.CUL  
-# │   └── WHCER048.ECO
-# └── Wheat/            # Contains experiment input and output files
-#     ├── SWSW7501.WHT
-#     ├── OVERVIEW.OUT
-#     └── PlantGro.OUT
-# ```
-# 
-# These files mimic the structure of a DSSAT installation (`C:/DSSAT48/`) but are self-contained within this repository. Examples reference these files using **relative paths**, ensuring portability.
-# 
-# ### 1.2. How the Provided DSSAT Files Were Generated
-# 
-# If you wish to reproduce the DSSAT files yourself, follow these steps:
-# 
-# 1. Launch DSSAT.
-# 2. Click “Selector”.
-# 3. Expand “Crops” and select “Wheat”.
-# 4. In the “Data” panel select the “SWSW7501.WHX” experiment.
-# 5. Click “Run” button in the toolbar.
-# 6. In the “Simulation” popup window, choose “CERES” as the crop model.
-# 7. Click “Run Model” and wait for the simulation to finish.
-# 
-# 
-# [![](https://markdown-videos-api.jorgenkh.no/youtube/dzKpvJSEXZc?vq)](https://www.youtube.com/embed/dzKpvJSEXZc?vq=hd1080)
-# 
-# Click to watch the steps!
-
-# ### 1.2 Load the *dpest* library or import each module
-
-# In[1]:
-
-
-#from dpestool.wheat.ceres import cul, eco
-# from dpestool.wheat import overview, plantgro
-# from dpestool import pst
-# from dpestool.wheat.utils import uplantgro
-
-import dpest 
-
-
-# ## 2. Instructions for Creating the PEST Input Files
-# 
-# This section demonstrates how to use the `dpest` library functions to create the **PEST input files** with practical examples. For each function, you will find:
-# 
-# - Example code showing how to call the function and the expected output messages.
-# - Output files generated by running these examples. These files will appear in your **current working directory**.
-# 
-# Including these output files allows any user to:
-# 
-# - See what the expected outputs should look like.
-# - Compare their results with the provided files for validation.
-# - Better understand how `dpest` works in practice.
-# 
-# These examples can be used as templates for your workflows or as a reference to verify your installation and usage of `dpest`.
-# 
-# **Important:**  
-# All files generated by `dpest` (such as `PEST_CONTROL.pst`, `.TPL`, `.INS`, etc.) will appear in your current working directory. This is necessary because the PEST utility `pestchek.exe` expects both the control file and all referenced files to be in the directory where the command is run, and does not accept relative paths to files located in subdirectories. Saving all files in the current directory ensures that validation and further PEST operations work correctly.
-# 
[…]
