+++ tests/test_pst.py
@@ -1,6 +1,3 @@
-import dpest
-from pathlib import Path
-
 def test_pst_file_creation(tmp_path):
     """Test creation of a PEST control file (.PST) with all required inputs."""
 
@@ -17,49 +14,53 @@ def test_pst_file_creation(tmp_path):
     assert overview_file.exists(), f"Missing: {overview_file}"
     assert plantgro_file.exists(), f"Missing: {plantgro_file}"
 
-    # Create the output directory
-    output_dir = tmp_path / "output"
-    output_dir.mkdir(parents=True, exist_ok=True)
+    # Convert to string once
+    cul_file = str(cul_file)
+    eco_file = str(eco_file)
+    overview_file = str(overview_file)
+    plantgro_file = str(plantgro_file)
+    output_dir = str(tmp_path / "output")
+    Path(output_dir).mkdir(parents=True, exist_ok=True)
 
     # Step 1: Generate parameter dicts using cul/eco functions
-    cul_tpl_path, cultivar_parameters = dpest.wheat.ceres.cul(
+    cultivar_parameters, cul_tpl_path = dpest.wheat.ceres.cul(
         P='P1D, P5',
         G='G1, G2, G3',
         PHINT='PHINT',
         cultivar='MANITOU',
-        cul_file_path=str(cul_file),
-        output_path=str(output_dir)
+        cul_file_path=cul_file,
+        output_path=output_dir
     )
 
-    eco_tpl_path, ecotype_parameters = dpest.wheat.ceres.eco(
+    ecotype_parameters, eco_tpl_path = dpest.wheat.ceres.eco(
         PHEN='P1, P2FR1',
         VERN='VEFF',
         ecotype='CAWH01',
-        eco_file_path=str(eco_file),
-        output_path=str(output_dir)
+        eco_file_path=eco_file,
+        output_path=output_dir
     )
 
     # Step 2: Generate observations using overview and plantgro
     overview_obs, overview_ins_path = dpest.wheat.overview(
         treatment='164.0 KG N/HA IRRIG',
-        overview_file_path=str(overview_file),
-        output_path=str(output_dir)
+        overview_file_path=overview_file,
+        output_path=output_dir
     )
 
     plantgro_obs, plantgro_ins_path = dpest.wheat.plantgro(
         treatment='164.0 KG N/HA IRRIG',
         variables=['LAID', 'CWAD', 'T#AD'],
-        plantgro_file_path=str(plantgro_file),
-        output_path=str(output_dir)
+        plantgro_file_path=plantgro_file,
+        output_path=output_dir
     )
 
     # Step 3: Define model command and file pairs
     model_command = r'py "C:\pest18\run_dssat.py"'
     input_output_pairs = [
-        (str(cul_tpl_path), str(cul_file)),
-        (str(eco_tpl_path), str(eco_file)),
-        (str(overview_ins_path), str(overview_file)),
-        (str(plantgro_ins_path), str(plantgro_file))
+        (str(cul_tpl_path), cul_file),
+        (str(eco_tpl_path), eco_file),
+        (str(overview_ins_path), overview_file),
+        (str(plantgro_ins_path), plantgro_file)
     ]
 
     # Step 4: Create .PST file
@@ -69,12 +70,12 @@ def test_pst_file_creation(tmp_path):
         dataframe_observations=[overview_obs, plantgro_obs],
         model_comand_line=model_command,
         input_output_file_pairs=input_output_pairs,
-        output_path=str(output_dir),
+        output_path=output_dir,
         pst_filename="PEST_CONTROL.pst"
     )
 
     # Step 5: Validate .pst file creation
-    pst_file = output_dir / "PEST_CONTROL.pst"
[…]
