+++ tests/test_pst.py
@@ -1,3 +1,6 @@
+import dpest
+from pathlib import Path
+
 def test_pst_file_creation(tmp_path):
     """Test creation of a PEST control file (.PST) with all required inputs."""
 
@@ -14,13 +17,15 @@ def test_pst_file_creation(tmp_path):
     assert overview_file.exists(), f"Missing: {overview_file}"
     assert plantgro_file.exists(), f"Missing: {plantgro_file}"
 
-    # Convert to string once
+    # Convert paths to strings after checks
     cul_file = str(cul_file)
     eco_file = str(eco_file)
     overview_file = str(overview_file)
     plantgro_file = str(plantgro_file)
-    output_dir = str(tmp_path / "output")
-    Path(output_dir).mkdir(parents=True, exist_ok=True)
+
+    # Create the output directory
+    output_dir = tmp_path / "output"
+    output_dir.mkdir(parents=True, exist_ok=True)
 
     # Step 1: Generate parameter dicts using cul/eco functions
     cultivar_parameters, cul_tpl_path = dpest.wheat.ceres.cul(
@@ -29,7 +34,7 @@ def test_pst_file_creation(tmp_path):
         PHINT='PHINT',
         cultivar='MANITOU',
         cul_file_path=cul_file,
-        output_path=output_dir
+        output_path=str(output_dir)
     )
 
     ecotype_parameters, eco_tpl_path = dpest.wheat.ceres.eco(
@@ -37,28 +42,28 @@ def test_pst_file_creation(tmp_path):
         VERN='VEFF',
         ecotype='CAWH01',
         eco_file_path=eco_file,
-        output_path=output_dir
+        output_path=str(output_dir)
     )
 
     # Step 2: Generate observations using overview and plantgro
     overview_obs, overview_ins_path = dpest.wheat.overview(
         treatment='164.0 KG N/HA IRRIG',
         overview_file_path=overview_file,
-        output_path=output_dir
+        output_path=str(output_dir)
     )
 
     plantgro_obs, plantgro_ins_path = dpest.wheat.plantgro(
         treatment='164.0 KG N/HA IRRIG',
         variables=['LAID', 'CWAD', 'T#AD'],
         plantgro_file_path=plantgro_file,
-        output_path=output_dir
+        output_path=str(output_dir)
     )
 
     # Step 3: Define model command and file pairs
     model_command = r'py "C:\pest18\run_dssat.py"'
     input_output_pairs = [
         (str(cul_tpl_path), cul_file),
-        (str(eco_tpl_path), eco_file),
+        (str(eco_tpl_path), eco_file)),
         (str(overview_ins_path), overview_file),
         (str(plantgro_ins_path), plantgro_file)
     ]
@@ -70,12 +75,12 @@ def test_pst_file_creation(tmp_path):
         dataframe_observations=[overview_obs, plantgro_obs],
         model_comand_line=model_command,
         input_output_file_pairs=input_output_pairs,
-        output_path=output_dir,
+        output_path=str(output_dir),
         pst_filename="PEST_CONTROL.pst"
     )
 
     # Step 5: Validate .pst file creation
-    pst_file = Path(output_dir) / "PEST_CONTROL.pst"
+    pst_file = output_dir / "PEST_CONTROL.pst"
     assert pst_file.exists(), "PEST control file was not created."
 
     # Step 6: Confirm first line and key content
@@ -85,7 +90,7 @@ def test_pst_file_creation(tmp_path):
         content = ''.join(lines).lower()
         required_sections = [
             '* control data',
-            '* lsqr',
+            '* lsqr'
             '* parameter groups',
             '* parameter data',
             '* observation groups',
