+++ tests/test_cul.py
@@ -5,23 +5,27 @@ def test_cul_template_generation(tmp_path):
     """Test generation of cultivar template files."""
     # Setup paths
     repo_root = Path(__file__).parent.parent
-    input_file = repo_root / "tests/DSSAT48_data/Genotype/WHCER048.CUL"
+    cul_file = repo_root / "tests/DSSAT48_data/Genotype/WHCER048.CUL"
     output_dir = tmp_path / "output"
 
     # Ensure the input file exists
-    assert input_file.exists(), f"Input file not found: {input_file}"
+    assert cul_file.exists(), f"Input file not found: {input_file}"
 
     # Create the output directory
     output_dir.mkdir(parents=True, exist_ok=True)
 
+    # Convert paths to strings
+    cul_file = str(cul_file)
+    output_dir = str(output_dir)
+
     # Call the dpest.wheat.ceres.cul function
     result = dpest.wheat.ceres.cul(
-        P='P1D, P5',
-        G='G1, G2, G3',
-        PHINT='PHINT',
-        cultivar='MANITOU',
-        cul_file_path=str(input_file),
-        output_path=str(output_dir)
+        P = 'P1D, P5',
+        G = 'G1, G2, G3',
+        PHINT = 'PHINT',
+        cultivar = 'MANITOU',
+        cul_file_path = cul_file,
+        output_path = str(output_dir)
     )
 
     # 1. Validate result is not None
