+++ tests/test_overview.py
@@ -6,7 +6,7 @@ def test_overview_instruction_generation(tmp_path):
     """Test generation of instruction file and observations from OVERVIEW.OUT."""
     # Setup paths
     repo_root = Path(__file__).parent.parent
-    input_file = repo_root / "tests/DSSAT48_data/Wheat/OVERVIEW.OUT"
+    overview_file = repo_root / "tests/DSSAT48_data/Wheat/OVERVIEW.OUT"
     output_dir = tmp_path / "output"
 
     # Ensure the input file exists
@@ -15,10 +15,14 @@ def test_overview_instruction_generation(tmp_path):
     # Create the output directory
     output_dir.mkdir(parents=True, exist_ok=True)
 
+    # Convert paths to strings
+    overview_file = str(overview_file)
+    output_dir = str(output_dir)
+
     # Call the dpest.wheat.overview function
     result = dpest.wheat.overview(
         treatment='164.0 KG N/HA IRRIG',
-        overview_file_path=str(input_file),
+        overview_file_path=overview_file,
         output_path=str(output_dir)
     )
 
