+++ tests/test_utils.py
@@ -1,3 +1,4 @@
+from pathlib import Path
 from dpest.utils import *
 
 # Get repo root
