### 2025-05-19 #8188 @DS4Ag
Hello @julienmalard

Thank you for your helpful feedback. Please find our point-by-point responses below:

**1. Include link to documentation in the article**
We have included the link on Line 52.

**2. The repository does not seem to include automated tests.**
We have now included an automated test suite that covers all package functions in the [`tests` directory](https://github.com/DS4Ag/dpest/tree/main/tests). Additionally, we provide a Jupyter notebook example ([usage_example.ipynb](https://github.com/DS4Ag/dpest/blob/main/examples/wheat/ceres/usage_example.ipynb)) that demonstrates all package functions in a practical workflow and includes steps for using PEST’s “check” functions to validate the generated input files.

**3. Community guidelines seem to be missing**
We have added a summary of the community guidelines in the last section of the README file. Additionally, we have included detailed guidelines in the [CONTRIBUTING.md](https://github.com/DS4Ag/dpest/blob/main/CONTRIBUTING.md) document.

**4. Consider briefly citing similar DSSAT wrapper packages in different programming languages, even if they don't specifically support PEST files.**
We have added sentences from Line 40 to Line 43.

**5. The top menu bar of the documentation is a bit confusing (and redundant with the sidebar). In addition, the fact that dpest occurs below table of contents and basic usage in the sidebar but actually links to a header above the table of contents in the main page is a bit confusing at first.**
We have removed the navigation bar from the top of the page. Additionally, the sidebar menu no longer appears on the home page.

We hope these revisions address your comments. Please let us know if any further changes are needed.
