+++ README.md
@@ -1,6 +1,6 @@
-# dpest: Automating PEST Control File Creation for DSSAT Crop Model Calibration
+# dpest
 
-[![dpest](https://github.com/DS4Ag/dpest/blob/main/icon.svg)](https://dpest.readthedocs.io/en/latest/)
+<img src="https://github.com/DS4Ag/dpest/blob/main/graphics/icon.svg" width="180" alt="dpest"/>
 
 [![Python version](https://img.shields.io/badge/python-3.7+-blue.svg)](https://www.python.org/downloads/)
 [![License](https://img.shields.io/badge/license-GPL--3.0-green.svg)](https://www.gnu.org/licenses/gpl-3.0)
