### 2025-06-05 272c90d
requires-python >= 3.10
