### 2025-06-05 f313063
requires-python >= 3.10
