### 2025-06-06 #8188 @julienmalard
Thanks @elbeejay . I reran the tests and they now work. Would you consider the following?
* Adding code coverage
* Including pdm installation instructions in the README and in the documentation
