### 2025-06-06 95cd9f5
Developer and Tester Setup Instructions
