+++ .github/workflows/ci.yml
@@ -1,0 +1,29 @@
+name: CI with Coverage
+
+on: [push, pull_request]
+
+jobs:
+  test-and-coverage:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Set up Python 3.10
+        uses: actions/setup-python@v4
+        with:
+          python-version: "3.10"
+
+      - name: Install PDM and dependencies
+        run: |
+          pip install --user pdm
+          pdm install --dev
+
+      - name: Run tests with coverage
+        run: |
+          pdm run coverage run -m pytest
+          pdm run coverage xml
+
+      - name: Upload coverage to Codecov
+        uses: codecov/[email protected]
+        with:
+          file: ./coverage.xml
