+++ .github/workflows/ci.yml
@@ -1,7 +1,3 @@
-name: CI with Coverage
-
-on: [push, pull_request]
-
 jobs:
   test-and-coverage:
     runs-on: ubuntu-latest
@@ -24,6 +20,6 @@ jobs:
           pdm run coverage xml
 
       - name: Upload coverage to Codecov
-        uses: codecov/[email protected]
+        uses: codecov/codecov-action@v3
         with:
           file: ./coverage.xml
