+++ .github/workflows/ci.yml
@@ -26,7 +26,7 @@ jobs:
           pdm run coverage xml
 
       - name: Upload coverage reports to Codecov
-          uses: codecov/codecov-action@v5
+        uses: codecov/codecov-action@v5
           with:
             token: ${{ secrets.CODECOV_TOKEN }}
             slug: DS4Ag/dpest
