+++ tests/test_overview.py
@@ -1,8 +1,10 @@
 import dpest
 from pathlib import Path
 import pandas as pd
+import os
+import pytest
 
-def test_overview(tmp_path):
+def test_overview_basic(tmp_path):
     """Test generation of instruction file and observations from OVERVIEW.OUT."""
     # Setup paths
     repo_root = Path(__file__).parent.parent
@@ -50,3 +52,163 @@ def test_overview(tmp_path):
     # 7. Check that the DataFrame has the expected columns
     expected_columns = {'variable_name', 'value_measured', 'group'}
     assert expected_columns.issubset(df.columns), f"Missing expected columns in DataFrame: {expected_columns - set(df.columns)}"
+
+
+def test_overview_with_optional_parameters(tmp_path):
+    """Test with all optional parameters specified"""
+    repo_root = Path(__file__).parent.parent
+    overview_file = repo_root / "tests/DSSAT48_data/Wheat/OVERVIEW.OUT"
+
+    # Custom variables and classification
+    test_vars = ['Anthesis (DAP)', 'Maturity (DAP)']
+    test_classification = {'Anthesis (DAP)': 'phenology', 'Maturity (DAP)': 'phenology'}
+
+    result = dpest.wheat.overview(
+        treatment='164.0 KG N/HA IRRIG',
+        overview_file_path=str(overview_file),
+        output_path=str(tmp_path),
+        suffix='TRT1',
+        variables=test_vars,
+        variables_classification=test_classification,
+        overview_ins_first_line="pif #",
+        mrk='@',
+        smk='#'
+    )
+
+    df, ins_path = result
+    # Verify custom parameters
+    assert 'TRT1' in ins_path
+    assert set(df['variable']) == set(test_vars)
+    assert df['group'].unique().tolist() == ['phenology']
+
+
+def test_overview_missing_required_arguments():
+    """Test missing required arguments"""
+    with pytest.raises(ValueError) as excinfo:
+        dpest.wheat.overview(treatment=None, overview_file_path="dummy/path")
+    assert "The 'treatment' argument is required" in str(excinfo.value)
+
+
+def test_overview_invalid_suffix(tmp_path):
+    """Test invalid suffix values"""
+    repo_root = Path(__file__).parent.parent
+    overview_file = repo_root / "tests/DSSAT48_data/Wheat/OVERVIEW.OUT"
+
+    # Test non-string suffix
+    with pytest.raises(ValueError) as excinfo:
+        dpest.wheat.overview(
+            treatment='TREATMENT',
+            overview_file_path=str(overview_file),
+            suffix=123
+        )
+    assert "Suffix must be a string" in str(excinfo.value)
+
+    # Test invalid characters
+    with pytest.raises(ValueError) as excinfo:
+        dpest.wheat.overview(
+            treatment='TREATMENT',
+            overview_file_path=str(overview_file),
+            suffix="bad$"
+        )
+    assert "only contain letters and numbers" in str(excinfo.value)
+
+    # Test too long suffix
+    with pytest.raises(ValueError) as excinfo:
+        dpest.wheat.overview(
+            treatment='TREATMENT',
+            overview_file_path=str(overview_file),
+            suffix="LONGSUFFIX"
+        )
+    assert "at most 4 characters" in str(excinfo.value)
+
+
+def test_overview_file_not_found(tmp_path):
+    """Test with non-existent input file"""
+    with pytest.raises(FileNotFoundError):
+        dpest.wheat.overview(
+            treatment='TREATMENT',
+            overview_file_path="nonexistent/file.out",
+            output_path=str(tmp_path)
+        )
+
+
[…]
