+++ tests/test_overview.py
@@ -119,3 +119,180 @@ def test_overview_output_structure(tmp_path):
         content = f.read()
         # Check for truncated variable names with suffix
         assert any('_TEST!' in line for line in content.split('\n'))
+
+
+def test_overview_missing_required_arguments():
+    """Test missing required 'treatment' argument"""
+    with pytest.raises(ValueError) as excinfo:
+        dpest.wheat.overview(
+            treatment=None,
+            overview_file_path="dummy/path"
+        )
+    assert "The 'treatment' argument is required" in str(excinfo.value)
+
+
+@pytest.mark.parametrize("invalid_suffix, error_msg", [
+    (123, "Suffix must be a string"),
+    ("bad!", "only contain letters and numbers"),
+    ("LONGSUFFIX", "at most 4 characters")
+])
+def test_overview_invalid_suffix(tmp_path, invalid_suffix, error_msg):
+    """Test invalid suffix values"""
+    repo_root = Path(__file__).parent.parent
+    overview_file = repo_root / "tests/DSSAT48_data/Wheat/OVERVIEW.OUT"
+
+    with pytest.raises(ValueError) as excinfo:
+        dpest.wheat.overview(
+            treatment='164.0 KG N/HA IRRIG',
+            overview_file_path=str(overview_file),
+            output_path=str(tmp_path),
+            suffix=invalid_suffix
+        )
+    assert error_msg in str(excinfo.value)
+
+
+def test_overview_file_not_found(tmp_path):
+    """Test non-existent input file handling"""
+    with pytest.raises(FileNotFoundError):
+        dpest.wheat.overview(
+            treatment='164.0 KG N/HA IRRIG',
+            overview_file_path="nonexistent/file.out",
+            output_path=str(tmp_path)
+        )
+
+
+@pytest.mark.parametrize("mrk, smk, error_msg", [
+    ('a', '!', "Primary marker validation failed"),
+    ('!', '!', "Primary and secondary markers cannot match")
+])
+def test_overview_invalid_markers(tmp_path, mrk, smk, error_msg):
+    """Test invalid marker configurations"""
+    repo_root = Path(__file__).parent.parent
+    overview_file = repo_root / "tests/DSSAT48_data/Wheat/OVERVIEW.OUT"
+
+    with pytest.raises(ValueError) as excinfo:
+        dpest.wheat.overview(
+            treatment='164.0 KG N/HA IRRIG',
+            overview_file_path=str(overview_file),
+            output_path=str(tmp_path),
+            mrk=mrk,
+            smk=smk
+        )
+    assert error_msg in str(excinfo.value)
+
+
+def test_overview_variable_filtering(tmp_path):
+    """Test filtering with specific variables"""
+    repo_root = Path(__file__).parent.parent
+    overview_file = repo_root / "tests/DSSAT48_data/Wheat/OVERVIEW.OUT"
+
+    test_vars = ['Anthesis (DAP)', 'Product wt (kg dm/ha;no loss)']
+
+    result = dpest.wheat.overview(
+        treatment='164.0 KG N/HA IRRIG',
+        overview_file_path=str(overview_file),
+        output_path=str(tmp_path),
+        variables=test_vars
+    )
+
+    df, _ = result
+    expected_vars = {v.replace(' ', '')[:20] for v in test_vars}
+    assert all(var in df['variable_name'].values for var in expected_vars)
+
+
+def test_overview_nonexistent_treatment(tmp_path):
+    """Test handling of non-existent treatment"""
+    repo_root = Path(__file__).parent.parent
+    overview_file = repo_root / "tests/DSSAT48_data/Wheat/OVERVIEW.OUT"
+
+    with pytest.raises(ValueError) as excinfo:
[…]
