+++ tests/test_overview.py
@@ -121,14 +121,21 @@ def test_overview_output_structure(tmp_path):
         assert any('_TEST!' in line for line in content.split('\n'))
 
 
-def test_overview_missing_required_arguments():
+def test_overview_missing_required_arguments(capsys):
     """Test missing required 'treatment' argument"""
-    with pytest.raises(ValueError) as excinfo:
-        dpest.wheat.overview(
-            treatment=None,
-            overview_file_path="dummy/path"
-        )
-    assert "The 'treatment' argument is required" in str(excinfo.value)
+    result = dpest.wheat.overview(
+        treatment=None,
+        overview_file_path="dummy/path"
+    )
+    captured = capsys.readouterr()
+    assert "The 'treatment' argument is required" in captured.out
+    assert result is None
+
+@pytest.mark.parametrize("invalid_suffix, error_msg", [
+    (123, "Suffix must be a string"),
+    ("bad!", "only contain letters and numbers"),
+    ("LONGSUFFIX", "at most 4 characters")
+])
 
 
 @pytest.mark.parametrize("invalid_suffix, error_msg", [
@@ -136,49 +143,84 @@ def test_overview_missing_required_arguments():
     ("bad!", "only contain letters and numbers"),
     ("LONGSUFFIX", "at most 4 characters")
 ])
-def test_overview_invalid_suffix(tmp_path, invalid_suffix, error_msg):
+def test_overview_invalid_suffix(tmp_path, invalid_suffix, error_msg, capsys):
     """Test invalid suffix values"""
     repo_root = Path(__file__).parent.parent
     overview_file = repo_root / "tests/DSSAT48_data/Wheat/OVERVIEW.OUT"
 
-    with pytest.raises(ValueError) as excinfo:
-        dpest.wheat.overview(
-            treatment='164.0 KG N/HA IRRIG',
-            overview_file_path=str(overview_file),
-            output_path=str(tmp_path),
-            suffix=invalid_suffix
-        )
-    assert error_msg in str(excinfo.value)
+    result = dpest.wheat.overview(
+        treatment='164.0 KG N/HA IRRIG',
+        overview_file_path=str(overview_file),
+        output_path=str(tmp_path),
+        suffix=invalid_suffix
+    )
+    captured = capsys.readouterr()
+    assert error_msg in captured.out
+    assert result is None
 
 
-def test_overview_file_not_found(tmp_path):
+def test_overview_file_not_found(tmp_path, capsys):
     """Test non-existent input file handling"""
-    with pytest.raises(FileNotFoundError):
-        dpest.wheat.overview(
-            treatment='164.0 KG N/HA IRRIG',
-            overview_file_path="nonexistent/file.out",
-            output_path=str(tmp_path)
-        )
+    result = dpest.wheat.overview(
+        treatment='164.0 KG N/HA IRRIG',
+        overview_file_path="nonexistent/file.out",
+        output_path=str(tmp_path)
+    )
+    captured = capsys.readouterr()
+    assert "does not exist" in captured.out
+    assert result is None
 
 
 @pytest.mark.parametrize("mrk, smk, error_msg", [
     ('a', '!', "Primary marker validation failed"),
     ('!', '!', "Primary and secondary markers cannot match")
 ])
-def test_overview_invalid_markers(tmp_path, mrk, smk, error_msg):
+def test_overview_invalid_markers(tmp_path, mrk, smk, error_msg, capsys):
     """Test invalid marker configurations"""
     repo_root = Path(__file__).parent.parent
     overview_file = repo_root / "tests/DSSAT48_data/Wheat/OVERVIEW.OUT"
 
-    with pytest.raises(ValueError) as excinfo:
-        dpest.wheat.overview(
[…]
