+++ tests/test_overview.py
@@ -275,16 +275,17 @@ def test_overview_different_output_formats(tmp_path):
 
 def test_overview_missing_yaml_file(tmp_path, capsys):
     """Test handling of missing YAML arguments file"""
-    # Create dummy overview file path that doesn't exist
-    non_existent_file = tmp_path / "nonexistent.yml"
+    # Create dummy .OUT file path that doesn't exist
+    non_existent_file = tmp_path / "nonexistent.OUT"
 
-    with pytest.raises(FileNotFoundError) as excinfo:
-        dpest.wheat.overview(
-            treatment='164.0 KG N/HA IRRIG',
-            overview_file_path=str(non_existent_file),
-            output_path=str(tmp_path)
-        )
-    assert "YAML file not found" in str(excinfo.value)
+    result = dpest.wheat.overview(
+        treatment='164.0 KG N/HA IRRIG',
+        overview_file_path=str(non_existent_file),
+        output_path=str(tmp_path)
+    )
+    captured = capsys.readouterr()
+    assert "YAML file not found" in captured.out
+    assert result is None
 
 
 def test_overview_missing_treatment_argument(tmp_path, capsys):
@@ -292,13 +293,14 @@ def test_overview_missing_treatment_argument(tmp_path, capsys):
     repo_root = Path(__file__).parent.parent
     overview_file = repo_root / "tests/DSSAT48_data/Wheat/OVERVIEW.OUT"
 
-    with pytest.raises(ValueError) as excinfo:
-        dpest.wheat.overview(
-            treatment=None,
-            overview_file_path=str(overview_file),
-            output_path=str(tmp_path)
-        )
-    assert "The 'treatment' argument is required" in str(excinfo.value)
+    result = dpest.wheat.overview(
+        treatment=None,
+        overview_file_path=str(overview_file),
+        output_path=str(tmp_path)
+    )
+    captured = capsys.readouterr()
+    assert "The 'treatment' argument is required" in captured.out
+    assert result is None
 
 
 def test_overview_duplicate_markers(tmp_path, capsys):
@@ -306,15 +308,16 @@ def test_overview_duplicate_markers(tmp_path, capsys):
     repo_root = Path(__file__).parent.parent
     overview_file = repo_root / "tests/DSSAT48_data/Wheat/OVERVIEW.OUT"
 
-    with pytest.raises(ValueError) as excinfo:
-        dpest.wheat.overview(
-            treatment='164.0 KG N/HA IRRIG',
-            overview_file_path=str(overview_file),
-            output_path=str(tmp_path),
-            mrk='!',
-            smk='!'
-        )
-    assert "mrk and smk must be different" in str(excinfo.value)
+    result = dpest.wheat.overview(
+        treatment='164.0 KG N/HA IRRIG',
+        overview_file_path=str(overview_file),
+        output_path=str(tmp_path),
+        mrk='!',
+        smk='!'
+    )
+    captured = capsys.readouterr()
+    assert "mrk and smk must be different" in captured.out
+    assert result is None
 
 
 def test_overview_auto_ins_extension(tmp_path):
@@ -325,10 +328,30 @@ def test_overview_auto_ins_extension(tmp_path):
     result = dpest.wheat.overview(
         treatment='164.0 KG N/HA IRRIG',
         overview_file_path=str(overview_file),
-        output_path=str(tmp_path),
-        output_filename="TEST_OUTPUT"
+        output_path=str(tmp_path)
     )
-
     _, ins_path = result
     assert ins_path.endswith(".ins")
 
+
+def test_overview_unexpected_error(tmp_path, capsys, monkeypatch):
+    """Test handling of unexpected exceptions"""
[…]
