+++ tests/test_overview.py
@@ -1,7 +1,6 @@
 import dpest
 from pathlib import Path
 import pandas as pd
-import os
 import pytest
 
 def test_overview(tmp_path):
@@ -78,6 +77,105 @@ def test_overview_with_optional_parameters(tmp_path):
     assert 'TRT1' in ins_path
     assert {'Anthesis_DAP_TRT1', 'Maturity_DAP_TRT1'}.issubset(set(df['variable_name'].values))
 
+def test_overview_variable_filtering(tmp_path):
+    """Test filtering with specific variables"""
+    repo_root = Path(__file__).parent.parent
+    overview_file = repo_root / "tests/DSSAT48_data/Wheat/OVERVIEW.OUT"
+
+    test_vars = ['Anthesis (DAP)', 'Product wt (kg dm/ha;no loss)']
+
+    result = dpest.wheat.overview(
+        treatment='164.0 KG N/HA IRRIG',
+        overview_file_path=str(overview_file),
+        output_path=str(tmp_path),
+        variables=test_vars
+    )
+
+    df, _ = result
+    # Check for base variable names in formatted output
+    assert any('Anthesis_DAP' in name for name in df['variable_name'].values)
+    assert any('Productwtkgdmha' in name for name in df['variable_name'].values)
+
+def test_overview_full_parameters(tmp_path):
+    """Test all optional parameters together"""
+    repo_root = Path(__file__).parent.parent
+    overview_file = repo_root / "tests/DSSAT48_data/Wheat/OVERVIEW.OUT"
+
+    custom_vars = ['Anthesis (DAP)', 'Maturity (DAP)']
+    custom_classification = {'Anthesis (DAP)': 'phenology', 'Maturity (DAP)': 'phenology'}
+
+    result = dpest.wheat.overview(
+        treatment='164.0 KG N/HA IRRIG',
+        overview_file_path=str(overview_file),
+        output_path=str(tmp_path),
+        suffix='TEST',
+        variables=custom_vars,
+        variables_classification=custom_classification,
+        overview_ins_first_line="pif #",
+        mrk='@',
+        smk='%'
+    )
+
+    df, ins_path = result
+    assert {'variable_name', 'value_measured', 'group'}.issubset(df.columns)
+    assert len(df) == len(custom_vars)
+
+    with open(ins_path, 'r') as f:
+        content = f.read()
+        assert content.startswith('pif # @')
+        assert '@Anthesis (DAP)@ %Anthesis_DAP_TEST%' in content
+        assert '@Maturity (DAP)@ %Maturity_DAP_TEST%' in content
+
+
+def test_overview_missing_treatment_argument(tmp_path, capsys):
+    """Test missing treatment argument validation"""
+    repo_root = Path(__file__).parent.parent
+    overview_file = repo_root / "tests/DSSAT48_data/Wheat/OVERVIEW.OUT"
+
+    result = dpest.wheat.overview(
+        treatment=None,
+        overview_file_path=str(overview_file),
+        output_path=str(tmp_path)
+    )
+    captured = capsys.readouterr()
+    assert "ValueError: The 'treatment' argument is required and must be specified by the user." in captured.out
+    assert result is None
+
+
+def test_overview_special_characters_in_treatment(tmp_path, capsys):
+    """Test treatment names with special characters"""
+    repo_root = Path(__file__).parent.parent
+    overview_file = repo_root / "tests/DSSAT48_data/Wheat/OVERVIEW.OUT"
+
+    result = dpest.wheat.overview(
+        treatment='164.0 KG N/HA (IRRIGATED)',
+        overview_file_path=str(overview_file),
+        output_path=str(tmp_path)
+    )
+
+    if result is None:
+        captured = capsys.readouterr()
+        assert "No data found for treatment" in captured.out
+    else:
[…]
