+++ tests/test_plantgro.py
@@ -1,6 +1,7 @@
 import dpest
 from pathlib import Path
 import pandas as pd
+import pytest
 
 def test_overview(tmp_path):
     """Test generation of instruction file and observations from PlantGro.OUT."""
