+++ tests/test_pst.py
@@ -103,7 +103,6 @@ def test_pst(tmp_path):
             assert section in content, f"Missing section: {section}"
 
 
-#######################3
 def test_pst_missing_both_parameters(tmp_path, capsys):
     """Test error when both cultivar and ecotype parameters are missing"""
     # Setup valid observations and files
@@ -136,113 +135,28 @@ def test_pst_missing_both_parameters(tmp_path, capsys):
     assert not (tmp_path / "PEST_CONTROL.pst").exists()
 
 
-# def test_pst_invalid_cultivar_type(tmp_path, capsys):
-#     """Test error when cultivar_parameters is not a dict"""
-#     # Setup valid ecotype parameters
-#     repo_root = Path(__file__).parent.parent
-#     eco_file = repo_root / "tests/DSSAT48_data/Genotype/WHCER048.ECO"
-#     ecotype_params, _ = dpest.wheat.ceres.eco(
-#         PHEN='P1',
-#         VERN='VEFF',
-#         ecotype='TEST',
-#         eco_file_path=str(eco_file),
-#         output_path=str(tmp_path)
-#     )
-#
-#     # Call with invalid cultivar type
-#     dpest.pst(
-#         cultivar_parameters="invalid",
-#         ecotype_parameters=ecotype_params,
-#         dataframe_observations=[pd.DataFrame()],
-#         model_comand_line='dummy',
-#         input_output_file_pairs=[('dummy.tpl', 'dummy.eco')],
-#         output_path=str(tmp_path)
-#     )
-#
-#     captured = capsys.readouterr()
-#     assert "must be a dictionary" in captured.out
-#     assert not (tmp_path / "PEST_CONTROL.pst").exists()
+def test_pst_invalid_cultivar_type(tmp_path, capsys):
+    """Test invalid cultivar_parameters type"""
+    # Use valid ecotype parameters
+    repo_root = Path(__file__).parent.parent
+    eco_file = repo_root / "tests/DSSAT48_data/Genotype/WHCER048.ECO"
+    ecotype_params, _ = dpest.wheat.ceres.eco(
+        PHEN='P1, P2FR1',  # Valid parameters from existing test
+        VERN='VEFF',
+        ecotype='CAWH01',  # Valid existing ecotype
+        eco_file_path=str(eco_file),
+        output_path=str(tmp_path)
+    )
 
+    dpest.pst(
+        cultivar_parameters="invalid_string",  # Invalid type
+        ecotype_parameters=ecotype_params,
+        dataframe_observations=[pd.DataFrame()],
+        model_comand_line='dummy',
+        input_output_file_pairs=[('dummy.tpl', str(eco_file))],
+        output_path=str(tmp_path)
+    )
 
-# def test_pst_missing_cul_extension(tmp_path, capsys):
-#     """Test error when .CUL file is missing with cultivar params"""
-#     # Setup valid cultivar parameters
-#     repo_root = Path(__file__).parent.parent
-#     cul_file = repo_root / "tests/DSSAT48_data/Genotype/WHCER048.CUL"
-#     cultivar_params, _ = dpest.wheat.ceres.cul(
-#         P='P1D',
-#         G='G1',
-#         PHINT='PHINT',
-#         cultivar='TEST',
-#         cul_file_path=str(cul_file),
-#         output_path=str(tmp_path)
-#     )
-#
-#     # Use invalid file pairs without .CUL
-#     dpest.pst(
-#         cultivar_parameters=cultivar_params,
-#         dataframe_observations=[pd.DataFrame()],
-#         model_comand_line='dummy',
-#         input_output_file_pairs=[('dummy.tpl', 'invalid.txt')],
[…]
