+++ tests/test_pst.py
@@ -1,5 +1,7 @@
+mport pandas as pd
 import dpest
 from pathlib import Path
+import pytest
 
 def test_pst(tmp_path):
     """Test creation of a PEST control file (.PST) with all required inputs."""
