+++ tests/test_pst.py
@@ -1,4 +1,4 @@
-mport pandas as pd
+import pandas as pd
 import dpest
 from pathlib import Path
 import pytest
