+++ tests/test_pst.py
@@ -329,7 +329,6 @@ def test_pst_dataframe_observations_required(tmp_path, capsys):
 
 def test_pst_input_output_file_pairs_validation(tmp_path, capsys):
     """Test validation rules for input_output_file_pairs"""
-    # Setup valid parameters
     repo_root = Path(__file__).parent.parent
     cul_file = repo_root / "tests/DSSAT48_data/Genotype/WHCER048.CUL"
     cultivar_params, cul_tpl = dpest.wheat.ceres.cul(
@@ -341,7 +340,6 @@ def test_pst_input_output_file_pairs_validation(tmp_path, capsys):
         output_path=str(tmp_path)
     )
 
-    # Setup valid observations
     overview_file = repo_root / "tests/DSSAT48_data/Wheat/OVERVIEW.OUT"
     overview_obs, _ = dpest.wheat.overview(
         treatment='164.0 KG N/HA IRRIG',
@@ -349,31 +347,31 @@ def test_pst_input_output_file_pairs_validation(tmp_path, capsys):
         output_path=str(tmp_path)
     )
 
-    # Case 1: Pair with wrong number of elements
+    # Case 1: Pair with wrong number of elements (as strings)
     dpest.pst(
         cultivar_parameters=cultivar_params,
         dataframe_observations=[overview_obs],
         model_comand_line='dummy',
-        input_output_file_pairs=[(cul_tpl, cul_file, 'extra')],  # 3 elements
+        input_output_file_pairs=[(str(cul_tpl), str(cul_file), 'extra')],  # 3 elements, all strings
         output_path=str(tmp_path)
     )
     captured = capsys.readouterr()
     assert "must contain exactly two elements" in captured.out
     assert not (tmp_path / "PEST_CONTROL.pst").exists()
 
-    # Case 2: Invalid first element extension
+    # Case 2: Invalid first element extension (as strings)
     dpest.pst(
         cultivar_parameters=cultivar_params,
         dataframe_observations=[overview_obs],
         model_comand_line='dummy',
-        input_output_file_pairs=[('invalid.txt', cul_file)],  # .txt instead of .tpl/.ins
+        input_output_file_pairs=[('invalid.txt', str(cul_file))],  # .txt instead of .tpl/.ins
         output_path=str(tmp_path)
     )
     captured = capsys.readouterr()
     assert "must be a .tpl or .ins file" in captured.out
     assert not (tmp_path / "PEST_CONTROL.pst").exists()
 
-    # Case 3: Valid pairs
+    # Case 3: Valid pairs (as strings)
     valid_pairs = [
         (str(cul_tpl), str(cul_file)),
         ('valid.ins', 'plantgro.out')
@@ -385,6 +383,5 @@ def test_pst_input_output_file_pairs_validation(tmp_path, capsys):
         input_output_file_pairs=valid_pairs,
         output_path=str(tmp_path)
     )
-    # Verify successful creation
     pst_file = tmp_path / "PEST_CONTROL.pst"
     assert pst_file.exists(), "PST file not created with valid pairs"
