+++ tests/test_pst.py
@@ -325,90 +325,3 @@ def test_pst_dataframe_observations_required(tmp_path, capsys):
     captured = capsys.readouterr()
     assert "`dataframe_observations` must be provided." in captured.out
     assert not (tmp_path / "PEST_CONTROL.pst").exists()
-
-####################
-
-def test_pst_input_output_file_pairs_validation(tmp_path):
-    """Test validation of input/output file pair requirements"""
-    # Reuse setup from working test
-    repo_root = Path(__file__).parent.parent
-    cul_file = repo_root / "tests/DSSAT48_data/Genotype/WHCER048.CUL"
-    eco_file = repo_root / "tests/DSSAT48_data/Genotype/WHCER048.ECO"
-    overview_file = repo_root / "tests/DSSAT48_data/Wheat/OVERVIEW.OUT"
-    plantgro_file = repo_root / "tests/DSSAT48_data/Wheat/PlantGro.OUT"
-
-    # Generate required parameters and files
-    cultivar_parameters, cul_tpl_path = dpest.wheat.ceres.cul(
-        P='P1D, P5',
-        G='G1, G2, G3',
-        PHINT='PHINT',
-        cultivar='MANITOU',
-        cul_file_path=str(cul_file),
-        output_path=str(tmp_path)
-    )
-
-    # Create dummy valid.ins file for testing
-    valid_ins = tmp_path / "valid.ins"
-    valid_ins.touch()
-
-    # Generate observations with real files
-    overview_obs, overview_ins_path = dpest.wheat.overview(
-        treatment='164.0 KG N/HA IRRIG',
-        overview_file_path=str(overview_file),
-        output_path=str(tmp_path)
-    )
-
-    plantgro_obs, plantgro_ins_path = dpest.wheat.plantgro(
-        treatment='164.0 KG N/HA IRRIG',
-        variables=['LAID', 'CWAD', 'T#AD'],
-        plantgro_file_path=str(plantgro_file),
-        output_path=str(tmp_path)
-    )
-
-    # Test case 1: Pair with incorrect number of elements
-    with pytest.raises(ValueError) as excinfo:
-        dpest.pst(
-            cultivar_parameters=cultivar_parameters,
-            dataframe_observations=[overview_obs, plantgro_obs],
-            model_comand_line='dummy',
-            input_output_file_pairs=[
-                (str(cul_tpl_path), str(cul_file), 'extra'),  # Invalid 3 elements
-                (str(valid_ins), str(overview_file))  # Valid pair with existing files
-            ],
-            output_path=str(tmp_path)
-        )
-    assert "must contain exactly two elements" in str(excinfo.value)
-
-    # Test case 2: Invalid first element extension
-    with pytest.raises(ValueError) as excinfo:
-        dpest.pst(
-            cultivar_parameters=cultivar_parameters,
-            dataframe_observations=[overview_obs, plantgro_obs],
-            model_comand_line='dummy',
-            input_output_file_pairs=[
-                ('invalid.txt', str(cul_file)),  # Invalid extension
-                (str(valid_ins), str(overview_file))  # Valid pair with existing files
-            ],
-            output_path=str(tmp_path)
-        )
-    assert "must be a .tpl or .ins file" in str(excinfo.value)
-
-    # Test case 3: Valid pairs (should create PST file)
-    valid_pairs = [
-        (str(cul_tpl_path), str(cul_file)),
-        (str(overview_ins_path), str(overview_file)),
-        (str(plantgro_ins_path), str(plantgro_file))
-    ]
-
-    dpest.pst(
-        cultivar_parameters=cultivar_parameters,
-        dataframe_observations=[overview_obs, plantgro_obs],
[…]
