### 2025-06-07 #8188 @DS4Ag
Hi @julienmalard,
Thanks for your suggestions!
Code coverage is now included, and PDM installation instructions have been added to the README and documentation.
