### 2025-06-08 #8188 @julienmalard
Looks good! Would it be possible to improve test coverage for some key files, for instance:
* [uplantgro.py](https://app.codecov.io/gh/DS4Ag/dpest/blob/main/dpest%2Fwheat%2Futils%2Fuplantgro.py)
* [functions.py](https://app.codecov.io/gh/DS4Ag/dpest/blob/main/dpest%2Ffunctions.py)
* [rmv_lsqr.py](https://app.codecov.io/gh/DS4Ag/dpest/blob/main/dpest%2Futils%2Frmv_lsqr.py)
* [rmv_svd.py](https://app.codecov.io/gh/DS4Ag/dpest/blob/main/dpest%2Futils%2Frmv_svd.py)

Other files have low coverage but mostly due to uncovered error handling lines, which is not as critical.
