+++ tests/test_cul.py
@@ -1,5 +1,7 @@
 import dpest
 from pathlib import Path
+import pytest
+import re
 
 def test_cul(tmp_path):
     """Test generation of cultivar template files."""
@@ -52,3 +54,87 @@ def test_cul(tmp_path):
     # 7. Check that the dictionary has the expected nested keys
     expected_keys = {'parameters', 'minima_parameters', 'maxima_parameters', 'parameters_grouped'}
     assert expected_keys.issubset(params), f"Missing expected keys in params: {expected_keys - set(params)}"
+
+
+def test_cul_missing_arguments_file(capsys, tmp_path):
+    """Test printed error when arguments.yml is missing."""
+    # Temporarily rename arguments.yml if it exists
+    repo_root = Path(__file__).parent.parent
+    cul_file = repo_root / "tests/DSSAT48_data/Genotype/WHCER048.CUL"
+    output_dir = tmp_path / "output"
+    output_dir.mkdir(parents=True, exist_ok=True)
+    module_dir = Path(dpest.wheat.ceres.__file__).parent
+    yml_path = module_dir / "arguments.yml"
+    yml_backup = module_dir / "arguments.yml.bak"
+    if yml_path.exists():
+        yml_path.rename(yml_backup)
+    try:
+        result = dpest.wheat.ceres.cul(
+            P='P1D, P5',
+            G='G1, G2, G3',
+            PHINT='PHINT',
+            cultivar='MANITOU',
+            cul_file_path=str(cul_file),
+            output_path=str(output_dir)
+        )
+        captured = capsys.readouterr()
+        assert "FileNotFoundError: YAML file not found:" in captured.out
+        assert result is None
+    finally:
+        # Restore the yaml file if it was renamed
+        if yml_backup.exists():
+            yml_backup.rename(yml_path)
+
+def test_cul_missing_cultivar(capsys, tmp_path):
+    """Test printed error when cultivar is not provided."""
+    repo_root = Path(__file__).parent.parent
+    cul_file = repo_root / "tests/DSSAT48_data/Genotype/WHCER048.CUL"
+    output_dir = tmp_path / "output"
+    output_dir.mkdir(parents=True, exist_ok=True)
+    result = dpest.wheat.ceres.cul(
+        P='P1D, P5',
+        G='G1, G2, G3',
+        PHINT='PHINT',
+        cul_file_path=str(cul_file),
+        output_path=str(output_dir)
+        # cultivar is omitted!
+    )
+    captured = capsys.readouterr()
+    assert "ValueError: The 'cultivar' argument is required and must be specified by the user." in captured.out
+    assert result is None
+
+def test_cul_invalid_cultivar(capsys, tmp_path):
+    """Test printed error when cultivar does not exist in file."""
+    repo_root = Path(__file__).parent.parent
+    cul_file = repo_root / "tests/DSSAT48_data/Genotype/WHCER048.CUL"
+    output_dir = tmp_path / "output"
+    output_dir.mkdir(parents=True, exist_ok=True)
+    result = dpest.wheat.ceres.cul(
+        P='P1D, P5',
+        G='G1, G2, G3',
+        PHINT='PHINT',
+        cultivar='INVALID',
+        cul_file_path=str(cul_file),
+        output_path=str(output_dir)
+    )
+    captured = capsys.readouterr()
+    assert "ValueError: The cultivar INVALID wasn't founded on file" in captured.out
+    assert result is None
+
+def test_cul_invalid_parameters(capsys, tmp_path):
+    """Test printed error when parameter does not exist in header."""
+    repo_root = Path(__file__).parent.parent
[…]
