+++ tests/test_eco.py
@@ -52,3 +52,135 @@ def test_eco(tmp_path):
     expected_keys = {'parameters', 'minima_parameters', 'maxima_parameters', 'parameters_grouped'}
     assert expected_keys.issubset(params), f"Missing expected keys in params: {expected_keys - set(params)}"
 
+def test_eco_missing_arguments_file(capsys, tmp_path):
+    """Test printed error when arguments.yml is missing."""
+    # Temporarily rename arguments.yml if it exists
+    repo_root = Path(__file__).parent.parent
+    eco_file = repo_root / "tests/DSSAT48_data/Genotype/WHCER048.ECO"
+    output_dir = tmp_path / "output"
+    output_dir.mkdir(parents=True, exist_ok=True)
+    module_dir = Path(dpest.wheat.ceres.__file__).parent
+    yml_path = module_dir / "arguments.yml"
+    yml_backup = module_dir / "arguments.yml.bak"
+    if yml_path.exists():
+        yml_path.rename(yml_backup)
+    try:
+        result = dpest.wheat.ceres.eco(
+            PHEN='P1, P2FR1',
+            VERN='VEFF',
+            ecotype='CAWH01',
+            eco_file_path=str(eco_file),
+            output_path=str(output_dir)
+        )
+        captured = capsys.readouterr()
+        assert "FileNotFoundError: YAML file not found:" in captured.out
+        assert result is None
+    finally:
+        # Restore the yaml file if it was renamed
+        if yml_backup.exists():
+            yml_backup.rename(yml_path)
+
+def test_eco_missing_ecotype(capsys, tmp_path):
+    """Test printed error when ecotype is not provided."""
+    repo_root = Path(__file__).parent.parent
+    eco_file = repo_root / "tests/DSSAT48_data/Genotype/WHCER048.ECO"
+    output_dir = tmp_path / "output"
+    output_dir.mkdir(parents=True, exist_ok=True)
+    result = dpest.wheat.ceres.eco(
+        PHEN='P1, P2FR1',
+        VERN='VEFF',
+        eco_file_path=str(eco_file),
+        output_path=str(output_dir)
+        # ecotype is omitted!
+    )
+    captured = capsys.readouterr()
+    assert "ValueError: The 'ecotype' argument is required and must be specified by the user." in captured.out
+    assert result is None
+
+def test_eco_invalid_ecotype(capsys, tmp_path):
+    """Test printed error when ecotype does not exist in file."""
+    repo_root = Path(__file__).parent.parent
+    eco_file = repo_root / "tests/DSSAT48_data/Genotype/WHCER048.ECO"
+    output_dir = tmp_path / "output"
+    output_dir.mkdir(parents=True, exist_ok=True)
+    result = dpest.wheat.ceres.eco(
+        PHEN='P1, P2FR1',
+        VERN='VEFF',
+        ecotype='INVALID',
+        eco_file_path=str(eco_file),
+        output_path=str(output_dir)
+    )
+    captured = capsys.readouterr()
+    assert "ValueError: The ecotype INVALID wasn't found in file" in captured.out
+    assert result is None
+
+def test_eco_invalid_parameters(capsys, tmp_path):
+    """Test printed error when parameter does not exist in header."""
+    repo_root = Path(__file__).parent.parent
+    eco_file = repo_root / "tests/DSSAT48_data/Genotype/WHCER048.ECO"
+    output_dir = tmp_path / "output"
+    output_dir.mkdir(parents=True, exist_ok=True)
+    result = dpest.wheat.ceres.eco(
+        P='INVALID_PARAM',
+        PHEN='P1, P2FR1',
+        VERN='VEFF',
+        ecotype='CAWH01',
+        eco_file_path=str(eco_file),
[…]
