### 2025-06-08 2b00e12
utils complete test
