### 2025-06-08 848c229
utils complete test
