+++ tests/test_utils.py
@@ -12,6 +12,43 @@ rstfle(pst_file_path, "restart")
 # Set PESTMODE to “prediction”
 pestmode(pst_file_path, "prediction")
 
+
+def test_pestmode_error_handling(tmp_path, capsys):
+    """Test error handling for pestmode function"""
+    # 1. FileNotFoundError
+    missing_file = tmp_path / "missing.pst"
+    pestmode(str(missing_file), "estimation")
+    captured = capsys.readouterr()
+    assert "Error: File" in captured.out
+
+    # 2. Non-string input
+    test_file1 = tmp_path / "test1.pst"
+    test_file1.write_text("\n" * 3)
+    pestmode(str(test_file1), 123)
+    captured = capsys.readouterr()
+    assert "ValueError: PESTMODE must be a string" in captured.out
+
+    # 3. Invalid mode
+    test_file2 = tmp_path / "test2.pst"
+    test_file2.write_text("\n" * 3)
+    pestmode(str(test_file2), "invalid_mode")
+    captured = capsys.readouterr()
+    assert "ValueError: PESTMODE must be one of" in captured.out
+
+    # 4. Insufficient lines
+    short_file = tmp_path / "short.pst"
+    short_file.write_text("single line\n")
+    pestmode(str(short_file), "estimation")
+    captured = capsys.readouterr()
+    assert "IndexError: File has only 1 lines" in captured.out
+
+    # 5. Insufficient values on line
+    sparse_file = tmp_path / "sparse.pst"
+    sparse_file.write_text("\n\nsingle_value\n")
+    pestmode(str(sparse_file), "estimation")
+    captured = capsys.readouterr()
+    assert "ValueError: PESTMODE value not found" in captured.out
+
 # Set RLAMBDA1 to 5.0
 rlambda1(pst_file_path, 5.0)
 
@@ -27,11 +64,97 @@ phiredlam(pst_file_path, 0.03)
 # Set NUMLAM to 10
 numlam(pst_file_path, 10)
 
+
+def test_numlam_error_handling(tmp_path, capsys):
+    """Test error handling for numlam function"""
+    # 1. FileNotFoundError
+    missing_file = tmp_path / "missing.pst"
+    numlam(str(missing_file), 10)
+    captured = capsys.readouterr()
+    assert "Error: File" in captured.out
+
+    # 2. ValueError: non-integer input
+    test_file1 = tmp_path / "test1.pst"
+    test_file1.write_text("\n" * 6)
+    numlam(str(test_file1), "invalid")
+    captured = capsys.readouterr()
+    assert "ValueError: NUMLAM must be an integer" in captured.out
+
+    # 3. ValueError: zero value
+    test_file2 = tmp_path / "test2.pst"
+    test_file2.write_text("\n" * 6)
+    numlam(str(test_file2), 0)
+    captured = capsys.readouterr()
+    assert "ValueError: NUMLAM cannot be zero" in captured.out
+
+    # 4. IndexError: insufficient lines
+    short_file = tmp_path / "short.pst"
+    short_file.write_text("\n" * 5)  # Only 5 lines
+    numlam(str(short_file), 10)
+    captured = capsys.readouterr()
+    assert "IndexError: File has only 5 lines" in captured.out
+
+    # 5. ValueError: insufficient values on line
+    sparse_file = tmp_path / "sparse.pst"
+    lines = ["\n"] * 5 + ["1 2 3 4\n"]  # Only 4 values on line 6
+    sparse_file.write_text("".join(lines))
+    numlam(str(sparse_file), 10)
+    captured = capsys.readouterr()
+    assert "ValueError: NUMLAM position not found" in captured.out
+
+
[…]
