+++ tests/test_utils.py
@@ -9,9 +9,71 @@ pst_file_path = repo_root / "tests/dpest_out/PEST_CONTROL.pst"
 # Set RSTFLE to “restart”
 rstfle(pst_file_path, "restart")
 
+
+def test_rstfle_error_handling(tmp_path, capsys):
+    """Test error handling for rstfle function"""
+    # 1. FileNotFoundError
+    missing_file = tmp_path / "missing.pst"
+    rstfle(str(missing_file), "restart")
+    captured = capsys.readouterr()
+    assert "Error: The file" in captured.out
+
+    # 2. Non-string input
+    test_file1 = tmp_path / "test1.pst"
+    test_file1.write_text("\n" * 3)
+    rstfle(str(test_file1), 123)
+    captured = capsys.readouterr()
+    assert "ValueError: RSTFLE must be a string" in captured.out
+
+    # 3. Invalid string value
+    rstfle(str(test_file1), "invalid")
+    captured = capsys.readouterr()
+    assert "ValueError: RSTFLE must be either 'restart' or 'norestart'" in captured.out
+
+    # 4. Insufficient lines
+    short_file = tmp_path / "short.pst"
+    short_file.write_text("\n" * 2)  # Only 2 lines
+    rstfle(str(short_file), "restart")
+    captured = capsys.readouterr()
+    assert "IndexError: Expected at least 3 lines" in captured.out
+
+    # 5. Empty target line
+    empty_line_file = tmp_path / "empty_line.pst"
+    lines = ["* control\n", "\n", "\n"]  # Empty line 3
+    empty_line_file.write_text("".join(lines))
+    rstfle(str(empty_line_file), "restart")
+    captured = capsys.readouterr()
+    assert "ValueError: RSTFLE line not found or is empty" in captured.out
+
+
+def test_rstfle_success(tmp_path):
+    """Test successful RSTFLE updates"""
+    # Create test file
+    test_file = tmp_path / "valid.pst"
+    original_line = "  restart  param1  param2\n"
+    lines = ["* control\n", "dummy line\n", original_line]
+    test_file.write_text("".join(lines))
+
+    # Test both valid values with different cases
+    for value in ["RESTART", "norestart"]:
+        rstfle(str(test_file), value)
+
+        # Verify update
+        updated_lines = test_file.read_text().splitlines()
+        updated_values = updated_lines[2].split()
+
+        # Check RSTFLE value (first value, lowercase)
+        assert updated_values[0] == value.lower()
+
+        # Verify other values remain unchanged
+        assert updated_values[1:] == ["param1", "param2"]
+
+        # Check formatting preservation
+        assert updated_lines[2].startswith("  "), "Leading whitespace not preserved"
+        assert "  " in updated_lines[2], "Value spacing not preserved"
+
 # Set PESTMODE to “prediction”
 pestmode(pst_file_path, "prediction")
-
 
 def test_pestmode_error_handling(tmp_path, capsys):
     """Test error handling for pestmode function"""
@@ -52,18 +114,267 @@ def test_pestmode_error_handling(tmp_path, capsys):
 # Set RLAMBDA1 to 5.0
 rlambda1(pst_file_path, 5.0)
 
+
+def test_rlambda1_error_handling(tmp_path, capsys):
+    """Test error handling for rlambda1 function"""
+    # 1. FileNotFoundError
+    missing_file = tmp_path / "missing.pst"
+    rlambda1(str(missing_file), 5.0)
+    captured = capsys.readouterr()
+    assert "Error: File" in captured.out
+
+    # 2. ValueError: non-float input
+    test_file1 = tmp_path / "test1.pst"
[…]
