### 2025-06-09 ba3934c
utils complete test
