### 2025-06-09 e4abac9
utils complete test
