### 2025-06-11 DS4Ag/dpest#4 @aleaf
bug: Absolute template and instruction file paths written to PEST control file
Hi, thanks for the opportunity to review this software. I have followed the example workflow in ``examples/wheat/ceres/usage_example.ipynb,`` and have been able to run all of the steps, but I noticed that the absolute filepaths were written in the PEST control file for the template and instruction files:
```
* model input/output
<absolute path to dpest folder>/examples/wheat/ceres/WHCER048_CUL.TPL ./DSSAT48_data/Genotype/WHCER048.CUL
<absolute path to dpest folder>/examples/wheat/ceres/WHCER048_ECO.TPL ./DSSAT48_data/Genotype/WHCER048.ECO
<absolute path to dpest folder>/examples/wheat/ceres/OVERVIEW.ins ./DSSAT48_data/Wheat/OVERVIEW.OUT
<absolute path to dpest folder>/examples/wheat/ceres/PlantGro.ins ./DSSAT48_data/Wheat/PlantGro.OUT
```
To ensure portability (for example, executing the PEST run on another computer), and for security considerations, I would recommend writing relative paths for the instruction and template files (which in the case of the example, would be just the filenames themselves).
