### 2025-06-11 DS4Ag/dpest#5 @aleaf
Add 'ipykernel' dependency to pyproject.toml
Running the example notebook ``examples/wheat/ceres/usage_example.ipynb`` requires installation of ipykernel (in my case, ``pip install ipykernel`` was suffient to install related dependencies as well. Suggest adding ipykernel as an optional dependency to ``pyproject.toml``, similar to the [project.optional-dependencies](https://github.com/geopandas/geopandas/blob/1a1585ddfba8ef229ebdfc05b1c20ef3eb1e67a7/pyproject.toml#L32) section used by other projects.
