### 2025-06-11 DS4Ag/dpest#6 @aleaf
Documentation: Pathing issues with examples; forward run script in usage_example.ipynb
In the [Example: Calibrating DSSAT for Wheat (CERES Model)](https://dpest.readthedocs.io/en/latest/example.html#) and
in [usage_example.ipynb](https://github.com/DS4Ag/dpest/blob/main/examples/wheat/ceres/usage_example.ipynb) there are many absolute filepaths (i.e., starting in `C:`). Recommend switching these to relative filepaths (relative to the working directory for the example(s)). To facilitate reproducibility of the examples, theses filepaths should be consistent with project file structure (i.e., so that a user can simply clone the repository and run the example without having to make any edits). For example, the python script in "Calibrating DSSAT for Wheat (CERES Model)" example references ``C://DSSAT48``, but in the actual project, it looks like this folder is [a subfolder] called ``DSSAT48_data``.

For the final step in usage_example.ipynb, the python script for the forward model run is specified as ``C:\pest18\run_dssat.py``.  Searching the project, I don't see a ``run_dssat.py``, but was able to make (presumably) something similar from the example script in the "Calibrating DSSAT for Wheat (CERES Model)" (after fixing the pathing). Recommend generating the forward run script as part of this example, or including it so that the example Notebook runs "out of the box".

Furthermore, I'm not sure that the default "model command" of ``py "C:\pest18\run_dssat.py`` would be valid on most systems. After making ``run_dssat.py``, to get PEST to successfully launch the forward run script, I had to change my "model command" to ``python run_dssat.py``.

To make things extra clear for prospective users, consider adding a download button for the python script example in ['Multiple Treatments' tutorial](https://dpest.readthedocs.io/en/latest/example_multiple_trts.html), similar to [this one](https://doi-usgs.github.io/modflow-setup/10min.html#create-a-setup-script-and-configuration-file). Name the downloaded file ``run_dssat.py`` for consistency.
