### 2025-06-11 DS4Ag/dpest#7 @aleaf
Documentation/Testing: suggest adding example script(s) to tests
To overcome the issues mentioned in #6 and prevent them from reoccuring, I recommend executing all code in the examples as part of the test suite. This ensures that examples shown in the documentation are consistent with the current state of the code, and that they will work for a user. This could be done a number of ways; in [this example](https://github.com/DOI-USGS/python-for-hydrology.git), notebooks are copied to the docs folder (see ``conf.py``) and then executed by the [nbsphinx](https://nbsphinx.readthedocs.io/en/0.9.7/) extension, which renders them into html.  [test_notebooks.py](https://github.com/DOI-USGS/python-for-hydrology/blob/main/tests/test_notebooks.py) in this repository also shows how notebooks can be executed in the test suite.
