### 2025-06-12 DS4Ag/dpest#8 @aleaf
Consider changing to package name to be more descriptive
I understand that there is already a "dssat-pest" R package that includes some of the same functionality, but consider changing the name from "dpest" to be more descriptive of what the package does. "dpest" isn't very descriptive; if anything, since there is just a "d" in front of PEST, it might suggest that the software is a modified version of PEST, perhaps with a different algorithm. "python-dssat-pest" would be lengthy, but descriptive, and could be shortened to "pydssat_pest," "pydssatpest" or maybe even "dssatpest" in the import name (since that would already be in a python context).
