### 2025-06-12 DS4Ag/dpest#9 @aleaf
Documentation: put narrative examples at the top
Currently the docs have a brief intro paragraph, followed by the API reference and then the examples at the end. I would suggest starting with at least some or all of the text of the JOSS paper (to orient users and provide some context), followed by the examples to show how dpest is used and a typical workflow, with the API reference at the end. Consider putting the table of contents on the left, which seems to be more common in software docs.
