+++ docs/example.rst
@@ -47,7 +47,7 @@ The first step is to create the cultivar Template File (``.TPL``) for the ``MANI
         G = 'G1, G2, G3', 
         PHINT = 'PHINT',
         cultivar = 'MANITOU',
-        cul_file_path = 'C:/DSSAT48/Genotype/WHCER048.CUL'
+        cul_file_path = './DSSAT48/Genotype/WHCER048.CUL'
     )  
 
 After running this function:  
@@ -66,14 +66,14 @@ For this experiment, key end-of-season crop performance metrics and phenological
     # Create OVERVIEW observations INS file
     overview_observations, overview_ins_path = dpest.wheat.overview(
         treatment = '164.0 KG N/HA IRRIG',  # Treatment Name
-        overview_file_path = 'C:/DSSAT48/Wheat/OVERVIEW.OUT'  # Path to the OVERVIEW.OUT file
+        overview_file_path = './DSSAT48/Wheat/OVERVIEW.OUT'  # Path to the OVERVIEW.OUT file
     )
 
     # Create PlantGro observations INS file
     plantgro_observations, plantgro_ins_path = dpest.wheat.plantgro(
         treatment = '164.0 KG N/HA IRRIG',  # Treatment Name
         variables = ['LAID', 'CWAD', 'T#AD'],  # Variables to calibrate
-        plantgro_file_path = 'C:/DSSAT48/Wheat/PlantGro.OUT'  # Path to the PlantGro.OUT file
+        plantgro_file_path = './DSSAT48/Wheat/PlantGro.OUT'  # Path to the PlantGro.OUT file
     )
 
 After running these functions:
@@ -99,40 +99,69 @@ The following Python script provides an example of how to run the ``DSSAT CERES-
     import subprocess
     from dpest.wheat.utils import uplantgro
 
-    def build_path(*args):
-        """
-        Construct a file path from multiple arguments.
-        """
-        return os.path.join(*args)
+    # User-editable section for system DSSAT installation
+    dssat_install_dir = r'C:\DSSAT48'  # System DSSAT installation folder
+    dssat_exe = os.path.join(dssat_install_dir, 'DSCSM048.EXE')
+    control_file = os.path.join(dssat_install_dir, 'Wheat', 'DSSBatch.v48')
 
-    # Define DSSAT root directory and output folder
-    dssat_path = 'C://DSSAT48/'
-    output_directory = 'C://DSSAT48/Wheat/'
+    # Project data directory (relative to script location)
+    project_dir = os.path.dirname(os.path.abspath(__file__))
+    data_dir = os.path.join(project_dir, 'DSSAT48')
+    output_dir = os.path.join(data_dir, 'Wheat')
 
-    # Set the working directory to the output folder
-    os.chdir(output_directory)
+    # Change working directory to the output directory
+    os.chdir(output_dir)
 
-    # Build the command to run DSSAT
-    main_executable = build_path(dssat_path, 'DSCSM048.EXE')
+    # Build and run DSSAT command
     module = 'CSCER048'
     switch = 'B'
-    control_file = build_path(dssat_path, 'Wheat/DSSBatch.v48')
-
-    # Create and execute the command 
-    command_line = ' '.join([main_executable, module, switch, control_file])
+    command_line = f'"{dssat_exe}" {module} {switch} "{control_file}"'
     result = subprocess.run(command_line, shell=True, check=True, capture_output=True, text=True)
-
-    # Print DSSAT execution output
     print(result.stdout)
 
     # Use uplantgro from dpest.wheat.utils to extract and update data from PlantGro.OUT if needed
     uplantgro(
[…]
