+++ docs/example_multiple_trts.rst
@@ -62,7 +62,7 @@ The first step is to create a template file (`.TPL`) for the `MANITOU` cultivar,
         G='G1, G2, G3',
         PHINT='PHINT',
         cultivar='MANITOU',
-        cul_file_path='C:/DSSAT48/Genotype/WHCER048.CUL'
+        cul_file_path='./DSSAT48/Genotype/WHCER048.CUL'
     )
 
 After running this function:
@@ -85,19 +85,19 @@ Use the `overview()` function to generate instruction files for each treatment t
 
     overview_observations_trt1, overview_ins_path_trt1 = overview(
         treatment='82.0 KG N/HA IRRIG',
-        overview_file_path='C:/DSSAT48/Wheat/OVERVIEW.OUT',
+        overview_file_path='./DSSAT48/Wheat/OVERVIEW.OUT',
         suffix='TRT1'
     )
 
     overview_observations_trt2, overview_ins_path_trt2 = overview(
         treatment='123.0 KG N/HA IRRIG',
-        overview_file_path='C:/DSSAT48/Wheat/OVERVIEW.OUT',
+        overview_file_path='./DSSAT48/Wheat/OVERVIEW.OUT',
         suffix='TRT2'
     )
 
     overview_observations_trt3, overview_ins_path_trt3 = overview(
         treatment='164.0 KG N/HA IRRIG',
-        overview_file_path='C:/DSSAT48/Wheat/OVERVIEW.OUT',
+        overview_file_path='./DSSAT48/Wheat/OVERVIEW.OUT',
         suffix='TRT3'
     )
 
@@ -114,21 +114,21 @@ Use the `plantgro()` function to generate instruction files for each treatment t
 
     plantgro_observations_trt1, plantgro_ins_path_trt1 = plantgro(
         treatment='82.0 KG N/HA IRRIG',
-        plantgro_file_path='C:/DSSAT48/Wheat/PlantGro.OUT',
+        plantgro_file_path='./DSSAT48/Wheat/PlantGro.OUT',
         variables=['LAID', 'CWAD', 'T#AD'],
         suffix='TRT1'
     )
 
     plantgro_observations_trt2, plantgro_ins_path_trt2 = plantgro(
         treatment='123.0 KG N/HA IRRIG',
-        plantgro_file_path='C:/DSSAT48/Wheat/PlantGro.OUT',
+        plantgro_file_path='./DSSAT48/Wheat/PlantGro.OUT',
         variables=['LAID', 'CWAD', 'T#AD'],
         suffix='TRT2'
     )
 
     plantgro_observations_trt3, plantgro_ins_path_trt3 = plantgro(
         treatment='164.0 KG N/HA IRRIG',
-        plantgro_file_path='C:/DSSAT48/Wheat/PlantGro.OUT',
+        plantgro_file_path='./DSSAT48/Wheat/PlantGro.OUT',
         variables=['LAID', 'CWAD', 'T#AD'],
         suffix='TRT3'
     )
@@ -145,7 +145,7 @@ Note that both instruction files (`OVERVIEW.INS` and `PlantGro.INS`) will be cre
 
 After creating the ``template file`` and ``instruction files`` for calibrating the ``MANITOU`` wheat cultivar across multiple treatments, the next step is to generate the ``PEST control file (.PST)``. This file integrates all necessary components and guides the calibration process.
 
-The `.PST` file is created using the variables obtained in **2.2** and **2.3**. Additionally, you need to specify the `command-line instruction` to execute the DSSAT model for all treatments. For more information on running DSSAT from the command line, visit the the `DSSAT web documentation <https://dssat.net/tools/tools-for-power-users/>`_.
[…]
