+++ docs/index.rst
@@ -1,55 +1,185 @@
 Home
 =======================
 
+A Python package for automating PEST (Parameter Estimation) file creation for DSSAT crop model calibration using time series data.
+
 What is dpest?
-=============
+==============
 
-`dpest` is a Python package designed to automate the creation of `PEST (Parameter Estimation and Uncertainty Analysis)`_ control files for calibrating `DSSAT (Decision Support System for Agrotechnology Transfer)`_ crop models. Currently, `dpest` is capable of calibrating DSSAT wheat models only. It generates template files for cultivar and ecotype parameters, instruction files for `OVERVIEW.OUT` and `PlantGro.OUT`, and the main PEST control file. A utility module is also included to extend `PlantGro.OUT` files for complete time series compatibility.
+``dpest`` is a Python package designed to automate the creation of `PEST (Parameter Estimation and Uncertainty Analysis) <https://pesthomepage.org/>`_ control files for calibrating `DSSAT (Decision Support System for Agrotechnology Transfer) <https://dssat.net/>`_ crop models. Currently, ``dpest`` is capable of calibrating DSSAT wheat models only. It generates template files for cultivar and ecotype parameters, instruction files for ``OVERVIEW.OUT`` and ``PlantGro.OUT``, and the main PEST control file. A utility module is also included to extend ``PlantGro.OUT`` files for complete time series compatibility.
 
-.. _PEST (Parameter Estimation and Uncertainty Analysis): https://pesthomepage.org/
-.. _DSSAT (Decision Support System for Agrotechnology Transfer): https://dssat.net/
+Documentation
+-------------
 
-This documentation provides a complete reference for using `dpest`.
+For detailed usage instructions and module references, see the full documentation on: `dpest.readthedocs.io/en/latest <https://dpest.readthedocs.io/en/latest/>`_.
 
+Key Features
+============
+
+- Automated generation of ``PEST control files (.PST)``.
+- Compatibility with DSSAT wheat models.
+- Creation of ``PEST template files (.TPL)`` for cultivar and ecotype parameters.
+- Generation of ``PEST instruction files (.INS)`` for key DSSAT output files.
+- Utility module to extend ``PlantGro.OUT`` files for complete time series.
 
 Installation
+============
+
+``dpest`` can be installed via pip from `PyPI <https://pypi.org/project/dpest/>`_.
+
+::
+
+    pip install dpest
+
+Prerequisites
+=============
+
+1. **DSSAT Software:** `DSSAT (version 4.8) <https://dssat.net/>`_ must be installed on your system.
+2. **DSSAT Experiment Data:** You'll need a DSSAT experiment to work with. The example below uses the ``SWSW7501WH N RESPONSE`` wheat experiment.
+3. **PEST Software:** You’ll need to have `PEST (version 18) <https://pesthomepage.org/>`_ installed.
+
+How to Use dpest?
 =================
-`dpest` can be installed via pip from `PyPI <https://pypi.org/project/dpest/>`_.
 
-.. code-block:: bash
+Jupyter notebook example
+------------------------
 
-   pip install dpest
+We provide a Jupyter notebook example (`usage_example.ipynb <https://github.com/DS4Ag/dpest/blob/main/examples/wheat/ceres/usage_example.ipynb>`_) that demonstrates all package functions in a practical workflow and includes steps for using PEST’s “check” functions to validate the generated input files.
+
+Read the Docs
+-------------
+
+For a **detailed, step-by-step example**, please refer to the official documentation:
[…]
