+++ docs/sidebar.rst
@@ -1,0 +1,29 @@
+.. toctree::
+   :maxdepth: 1
+   :caption: Home
+
+   Home <self>
+
+.. toctree::
+   :maxdepth: 2
+   :caption: PEST control file
+
+   dpest.pst
+   utils
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Wheat (PEST input files)
+
+   dpest.wheat.ceres.cul
+   dpest.wheat.ceres.eco
+   dpest.wheat.overview
+   dpest.wheat.plantgro
+   dpest.wheat.utils.uplantgro
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Examples
+
+   example
+   example_multiple_trts
