+++ pyproject.toml
@@ -31,12 +31,9 @@ build-backend = "setuptools.build_meta"
 Homepage = "https://github.com/DS4Ag/dpest"
 Issues = "https://github.com/DS4Ag/dpest/issues"
 
-[project.optional-dependencies]
+[dependency-groups]Add commentMore actions
+
 dev = [
     "pytest>=8.4.0",
     "coverage>=7.0.0",
-]
-notebooks = [
-    "ipykernel>=6.29.5",
-    "jupyter>=7.4.3. "
 ]
