+++ docs/conf.py
@@ -56,7 +56,4 @@ html_theme_options = {
     "show_nav_level": 2,  # Ensure navigation is visible up to depth 2
     "navigation_with_keys": True,  # Enable keyboard navigation
     "navbar_center": [],
-    "toc": {
-        "root": "sidebar",
-    },
 }
