+++ docs/index.rst
@@ -1,4 +1,4 @@
-Home
+dpest
 =======================
 
 A Python package for automating PEST (Parameter Estimation) file creation for DSSAT crop model calibration using time series data.
