+++ docs/example_multiple_trts.rst
@@ -199,7 +199,7 @@ The following Python script provides an example of how to run the ``DSSAT CERES-
 
 **Download the example of a Python script to run DSSAT**
 
-`run_dssat_mtrt.py <https://github.com/DS4Ag/dpest/blob/main/examples/wheat/ceres/run_dssat_mtrt.py>`_ *(Click to download if not already in your directory)*
+`run_dssat_mtrt.py <https://raw.githubusercontent.com/DS4Ag/dpest/refs/heads/main/examples/wheat/ceres/run_dssat_mtrt.py>`_ *(Click to download if not already in your directory)*
 
 .. important::
 
