### 2025-06-13 b7006fc
Optional dependencies
