### 2025-06-13 DS4Ag/dpest#5 @DS4Ag
Thank you for the observation. We have added `ipykernel` and `notebook` as `project.optional-dependencies` in `thepyproject.toml` file. Additionally, we included the instructions on how to [run the example provided in the example notebook file](https://github.com/DS4Ag/dpest?tab=readme-ov-file#jupyter-notebook-example) in the README section.
