### 2025-06-13 f1e5570
Include notebook tests
