+++ .github/workflows/tests.yml
@@ -1,3 +1,27 @@
+#name: Tests
+#on: [push, pull_request]
+#
+#jobs:
+#  test:
+#    runs-on: ubuntu-latest
+#    steps:
+#    - uses: actions/checkout@v4
+#
+#    - name: Set up Python 3.10
+#      uses: actions/setup-python@v4
+#      with:
+#        python-version: "3.10"
+#
+#    - name: Install dependencies
+#      run: |
+#        python -m pip install --upgrade pip
+#        pip install dpest
+#        pip install pytest
+#
+#    - name: Run tests
+#      run: |
+#        pytest tests/ -v
+
 name: Tests
 on: [push, pull_request]
 
@@ -5,19 +29,19 @@ jobs:
   test:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v4
+      - uses: actions/checkout@v4
 
-    - name: Set up Python 3.10
-      uses: actions/setup-python@v4
-      with:
-        python-version: "3.10"
+      - name: Set up Python 3.10
+        uses: actions/setup-python@v4
+        with:
+          python-version: "3.10"
 
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install dpest
-        pip install pytest
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install dpest
+          pip install -e .[dev]
 
-    - name: Run tests
-      run: |
-        pytest tests/ -v
+      - name: Run unit tests
+        run: |
+          pytest tests/ -v
