### 2025-06-13 9b07db6
Include notebook tests
