+++ tests/test_uplantgro.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 
 def test_uplantgro_success(capsys):
-    plantgro_path = Path('tests/DSSAT48_data/Wheat/PlantGro.OUT').resolve()
+    plantgro_path = Path('tests/DSSAT48/Wheat/PlantGro.OUT').resolve()
 
     result = dpest.wheat.utils.uplantgro(
         str(plantgro_path),
@@ -22,7 +22,7 @@ def test_uplantgro_success(capsys):
 
 
 def test_invalid_treatment_type(capsys):
-    plantgro_path = Path('tests/DSSAT48_data/Wheat/PlantGro.OUT').resolve()
+    plantgro_path = Path('tests/DSSAT48/Wheat/PlantGro.OUT').resolve()
     result = dpest.wheat.utils.uplantgro(str(plantgro_path), 123, ['LAID'])
     captured = capsys.readouterr()
     assert "ValueError: The 'treatment' must be a non-empty string." in captured.out
@@ -30,7 +30,7 @@ def test_invalid_treatment_type(capsys):
 
 
 def test_invalid_variables_type(capsys):
-    plantgro_path = Path('tests/DSSAT48_data/Wheat/PlantGro.OUT').resolve()
+    plantgro_path = Path('tests/DSSAT48/Wheat/PlantGro.OUT').resolve()
     result = dpest.wheat.utils.uplantgro(str(plantgro_path), 'Valid Treatment', None)
     captured = capsys.readouterr()
     assert "ValueError: The 'variables' should be a non-empty string or a list of strings." in captured.out
@@ -38,7 +38,7 @@ def test_invalid_variables_type(capsys):
 
 
 def test_invalid_nspaces_parameters(capsys):
-    plantgro_path = Path('tests/DSSAT48_data/Wheat/PlantGro.OUT').resolve()
+    plantgro_path = Path('tests/DSSAT48/Wheat/PlantGro.OUT').resolve()
 
     # Test all numeric parameters in one function
     for param in ['nspaces_year_header', 'nspaces_doy_header', 'nspaces_columns_header']:
@@ -54,7 +54,7 @@ def test_invalid_nspaces_parameters(capsys):
 
 
 def test_file_not_found(capsys):
-    non_existent_path = Path('tests/DSSAT48_data/Wheat/NonExistentFile.OUT')
+    non_existent_path = Path('tests/DSSAT48/Wheat/NonExistentFile.OUT')
     result = dpest.wheat.utils.uplantgro(str(non_existent_path), 'Valid Treatment', ['LAID'])
     captured = capsys.readouterr()
     assert "FileNotFoundError: The file 'NonExistentFile.OUT' does not exist" in captured.out
