+++ .github/workflows/tests.yml
@@ -40,7 +40,6 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install dpest
-          pip install pytest
           pip install -e .[dev]
 
       - name: Run unit tests
