+++ .github/workflows/ci.yml
@@ -20,10 +20,28 @@ jobs:
           pip install --user pdm
           pdm install --dev
 
+#      - name: Run tests with coverage
+#        run: |
+#          pdm run coverage run -m pytest
+#          pdm run coverage xml
+
+#      - name: Run tests with coverage
+#        env:
+#          COVERAGE_PROCESS_START: ${{ github.workspace }}/.coveragerc
+#        run: |
+#          pdm run coverage run -m pytest tests/
+#          pdm run coverage combine  # Merge subprocess coverage files
+#          pdm run coverage xml
+
       - name: Run tests with coverage
+        env:
+          COVERAGE_PROCESS_START: ${{ github.workspace }}/.coveragerc
         run: |
-          pdm run coverage run -m pytest
-          pdm run coverage xml
+          # Explicitly run notebook tests and unit tests
+          pdm run python -m coverage run -m pytest tests/test_notebooks.py
+          pdm run python -m coverage run -m pytest tests/
+          pdm run python -m coverage combine
+          pdm run python -m coverage xml
 
       - name: Upload coverage reports to Codecov
         uses: codecov/codecov-action@v5
