+++ .github/workflows/ci.yml
@@ -20,19 +20,6 @@ jobs:
           pip install --user pdm
           pdm install --dev
 
-#      - name: Run tests with coverage
-#        run: |
-#          pdm run coverage run -m pytest
-#          pdm run coverage xml
-
-#      - name: Run tests with coverage
-#        env:
-#          COVERAGE_PROCESS_START: ${{ github.workspace }}/.coveragerc
-#        run: |
-#          pdm run coverage run -m pytest tests/
-#          pdm run coverage combine  # Merge subprocess coverage files
-#          pdm run coverage xml
-
       - name: Run tests with coverage
         env:
           COVERAGE_PROCESS_START: ${{ github.workspace }}/.coveragerc
