+++ .github/workflows/ci.yml
@@ -33,6 +33,5 @@ jobs:
       - name: Upload coverage reports to Codecov
         uses: codecov/codecov-action@v5
         with:
-          # token: ${{ secrets.CODECOV_TOKEN }}  # Optional for public repos
           slug: DS4Ag/dpest
           files: ./coverage.xml
