### 2025-06-13 DS4Ag/dpest#9 @DS4Ag
Thank you for the suggestion!
We have moved the "Table of Contents" from the main page body to the left sidebar. Additionally, we have added the a dpest Introduction, Key Features, and a Basic Usage example on the main page.
