### 2025-06-14 606a656
pyEMU citation
