+++ paper/paper.bib
@@ -57,3 +57,17 @@
     url = {https://scisoc.confex.com/scisoc/2024am/meetingapp.cgi/Paper/161215}
 }
 
+@article{White:2016,
+    title = {A python framework for environmental model uncertainty analysis},
+    journal = {Environmental Modelling & Software},
+    volume = {85},
+    pages = {217-228},
+    year = {2016},
+    issn = {1364-8152},
+    doi = {https://doi.org/10.1016/j.envsoft.2016.08.017},
+    url = {https://www.sciencedirect.com/science/article/pii/S1364815216305461},
+    author = {Jeremy T. White and Michael N. Fienen and John E. Doherty},
+    keywords = {Uncertainty analysis, FOSM, Python, Model independent, Highly-parameterized, Data worth analysis},
+    abstract = {We have developed pyEMU, a python framework for Environmental Modeling Uncertainty analyses, open-source tool that is non-intrusive, easy-to-use, computationally efficient, and scalable to highly-parameterized inverse problems. The framework implements several types of linear (first-order, second-moment (FOSM)) and non-linear uncertainty analyses. The FOSM-based analyses can also be completed prior to parameter estimation to help inform important modeling decisions, such as parameterization and objective function formulation. Complete workflows for several types of FOSM-based and non-linear analyses are documented in example notebooks implemented using Jupyter that are available in the online pyEMU repository. Example workflows include basic parameter and forecast analyses, data worth analyses, and error-variance analyses, as well as usage of parameter ensemble generation and management capabilities. These workflows document the necessary steps and provides insights into the results, with the goal of educating users not only in how to apply pyEMU, but also in the underlying theory of applied uncertainty quantification.}
+}
+
