### 2025-06-14 bac8bad
pyEMU citation
