### 2025-06-14 #8188 @DS4Ag
Hi @aleaf,

Thank you for your helpful feedback!

We have addressed comments https://github.com/DS4Ag/dpest/issues/4, https://github.com/DS4Ag/dpest/issues/5, https://github.com/DS4Ag/dpest/issues/6, https://github.com/DS4Ag/dpest/issues/7, https://github.com/DS4Ag/dpest/issues/8, https://github.com/DS4Ag/dpest/issues/9.

We agree with your assessment of [pyEMU](https://github.com/pypest/pyemu) capabilities. For that reason, it is the [dpest.pst](https://github.com/DS4Ag/dpest/blob/main/dpest/pst.py) function backbone. Currently, the `pst()` function takes the data in form of directories from the `cul()` and `eco()` functions and dataframes from the `overview()` and `plantgro()` functions and uses `pyEMU` to generate the `.PST` file. However, the  `dpest.utils` functions modify some `.PST` file parameters directly in the file without using `pyEMU`, because it reconstructs the entire file from scratch using the current state of its objects, instead of updating lines in the original file.

Additionally, since `dpest` uses `pyEMU` to create the `PST` file, it is structurally compatible with `PEST++`. However, DSSAT’s fixed-width requirements for Cultivar parameters conflict with PEST++'s requirement to write as many significant figures to a parameter space as they can. For that reason, we only mention `PEST` for calibration using `dpest` generated files, instead of `PEST/ PEST++`.

Changes in the paper:

- Lines 43-46: included `pyEMU` citation
- Line 15-15:  We prefer to keep the original phrasing since this is the first mention of PEST in the manuscript, and we want to introduce the full name before using the acronym.
- Line 17: We deleted the "high computational speed" phrase since we do not have any data we can use to prove it.

Please let us know if any further changes are needed.

Luis
