### 2025-06-19 210fdc4
Improved Data Filtering in filter_dataframe() for uplantgro()
