+++ paper/paper.md
@@ -27,7 +27,7 @@ license: GPLv3+
 
 Process-based crop models simulate plant growth and can support crop improvement research by enacting different “what-if” scenarios. Decision Support System for Agrotechnology Transfer (DSSAT) is one of the most commonly used crop modeling platforms, containing modules that can be used to model different species [@JONES:2003]. To simulate different crop varieties, DSSAT calibration is needed. This involves adjusting model parameters, which are values stored in a model input file, to determine which sets of values give rise to simulations that most closely match measured data from field experiments; these data can include single timepoint measurements as well as time-series information. 
 
-While the DSSAT installation includes Graphical User Interface (GUI) tools specifically designed for calibration, they are not amenable to time-series data. In contrast, the Parameter Estimation and Uncertainty Analysis (PEST), a Command Line Interface (CLI) software, is a model-independent tool [@Doherty:2015] that can calibrate DSSAT models using not only the end-of-season crop data but also measurements collected within-season. To carry out model calibration, PEST requires three types of input files: template files (.TPL) that specify the parameters to calibrate, instruction files (.INS) that contain the guidelines to extract the model outputs, and the control file (.PST) that includes the specifications to manage all the calibration settings. Generation of these input files for PEST-based DSSAT calibration, however, can be a complex and time-consuming process, making it difficult to scale, for example, across many cultivars or strains.   
+While the DSSAT installation includes Graphical User Interface (GUI) tools specifically designed for calibration, they are not amenable to time-series data. In contrast, the PEST software suite for parameter estimation and uncertainty analysis [@Doherty:2015], a command-line interface (CLI) tool, is a model-independent tool that can calibrate DSSAT models using not only the end-of-season crop data but also measurements collected within-season. To carry out model calibration, PEST requires three types of input files: template files (.TPL) that specify the parameters to calibrate, instruction files (.INS) that contain the guidelines to extract the model outputs, and the control file (.PST) that includes the specifications to manage all the calibration settings. Generation of these input files for PEST-based DSSAT calibration, however, can be a complex and time-consuming process, making it difficult to scale, for example, across many cultivars or strains.   
 
 To address these challenges, we developed *dpest*, a Python library that streamlines the generation of PEST input files for calibrating DSSAT; the current version supports DSSAT’s wheat models. Importantly, using dpest enables researchers to script the entire calibration process, thereby enhancing the efficiency of the workflow to scale easily across multiple varieties. 
 
