### 2025-06-25 #8188 @elbeejay
@DS4Ag can you make a few small edits?

- The title and author list for the zenodo archive needs to match the JOSS paper exactly
- The bot looks to be correct, `10.1111/gwat.12360` (https://doi.org/10.1111/gwat.12360) looks like a valid DOI for the "Calibration and Uncertainty Analysis for Complex Environmental Models" entry
- If you could edit the .bib entry for "A python framework for environmental model uncertainty analysis" to just `10.1016/j.envsoft.2016.08.017` I think that would resolve that "invalid DOI" error above

Otherwise the paper and materials look good, so let me know when you've had a chance to make those changes.
