### 2025-06-26 #8188 @DS4Ag
Hi @elbeejay,

I've made the following updates:

1. **Zenodo metadata** now matches the JOSS paper exactly (title, authors, license).
2. Fixed the DOI for White et al. (2016) in the `.bib` file to `10.1016/j.envsoft.2016.08.017`.
3. Confirmed the Doherty (2015) book citation is correct without a DOI, as `10.1111/gwat.12360` refers to a book review, not the original work.

All requested changes are complete. Thank you for your guidance!

Best regards,
Luis
