+++ paper/paper.bib
@@ -64,7 +64,7 @@
     pages = {217-228},
     year = {2016},
     issn = {1364-8152},
-    doi = {https://doi.org/10.1016/j.envsoft.2016.08.017},
+    doi = {10.1016/j.envsoft.2016.08.017},
     url = {https://www.sciencedirect.com/science/article/pii/S1364815216305461},
     author = {Jeremy T. White and Michael N. Fienen and John E. Doherty},
     keywords = {Uncertainty analysis, FOSM, Python, Model independent, Highly-parameterized, Data worth analysis},
