### 2025-07-01 #8188 @elbeejay
Great thanks @DS4Ag - I appreciate the clarification on the book review. If you could remove the parenthetical "(Contact person)" in the Zenodo entry I think that would be our preference, but given everything else looks good, I'll recommend this for publication.
