### 2025-07-01 #8188 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
